Struct devices::bat::GoldfishBattery

pub struct GoldfishBattery {
    state: Arc<Mutex<GoldfishBatteryState>>,
    mmio_base: u32,
    irq_num: u32,
    irq_evt: IrqLevelEvent,
    activated: bool,
    monitor_thread: Option<JoinHandle<()>>,
    kill_evt: Option<Event>,
    tube: Option<Tube>,
    create_power_monitor: Option<Box<dyn CreatePowerMonitorFn>>,
}

GoldFish Battery state

Fields
state: Arc<Mutex<GoldfishBatteryState>>mmio_base: u32irq_num: u32irq_evt: IrqLevelEventactivated: boolmonitor_thread: Option<JoinHandle<()>>kill_evt: Option<Event>tube: Option<Tube>create_power_monitor: Option<Box<dyn CreatePowerMonitorFn>>
Implementations

impl GoldfishBattery

pub fn new(
mmio_base: u64,
irq_num: u32,
irq_evt: IrqLevelEvent,
tube: Tube,
create_power_monitor: Option<Box<dyn CreatePowerMonitorFn>>
) -> Result<Self, BatteryError>

Create GoldfishBattery device model

mmio_base - The 32-bit mmio base address.

irq_num - The corresponding interrupt number of the irq_evt which will be put into the ACPI DSDT.

irq_evt - The interrupt event used to notify driver about the battery properties changing.

socket - Battery control socket

pub fn keep_rds(&self) -> Vec<RawDescriptor>

return the descriptors used by this device

fn start_monitor(&mut self)

start a monitor thread to monitor the events from host