Trait devices::bus::BusDevice

pub trait BusDevice: Send {
    fn debug_label(&self) -> String;

    fn device_id(&self) -> DeviceId;

    fn read(&mut self, offset: BusAccessInfo, data: &mut [u8]) { ... }

    fn write(&mut self, offset: BusAccessInfo, data: &[u8]) { ... }

    fn config_register_write(
        &mut self, 
        reg_idx: usize, 
        offset: u64, 
        data: &[u8]
    ) -> ConfigWriteResult { ... }

    fn config_register_read(&self, reg_idx: usize) -> u32 { ... }

    fn virtual_config_register_write(&mut self, reg_idx: usize, value: u32) { ... }

    fn virtual_config_register_read(&self, reg_idx: usize) -> u32 { ... }

    fn on_sandboxed(&mut self) { ... }

    fn get_ranges(&self) -> Vec<(BusRange, BusType)> { ... }

    fn destroy_device(&mut self) { ... }

    fn is_bridge(&self) -> Option<u8> { ... }
}

Trait for devices that respond to reads or writes in an arbitrary address space.

The device does not care where it exists in address space as each method is only given an offset into its allocated portion of address space.

Required methods

fn debug_label(&self) -> String

Returns a label suitable for debug output.

fn device_id(&self) -> DeviceId

Returns a unique id per device type suitable for metrics gathering.

Provided methods

fn read(&mut self, offset: BusAccessInfo, data: &mut [u8])

Reads at offset from this device

fn write(&mut self, offset: BusAccessInfo, data: &[u8])

Writes at offset into this device

fn config_register_write(
&mut self,
reg_idx: usize,
offset: u64,
data: &[u8]
) -> ConfigWriteResult

Sets a register in the configuration space. Only used by PCI.

reg_idx - The index of the config register to modify.

offset - Offset in to the register.

fn config_register_read(&self, reg_idx: usize) -> u32

Gets a register from the configuration space. Only used by PCI.

reg_idx - The index of the config register to read.

fn virtual_config_register_write(&mut self, reg_idx: usize, value: u32)

Sets a register in the virtual config space. Only used by PCI.

reg_idx - The index of the config register to modify.

value - The value to be written.

fn virtual_config_register_read(&self, reg_idx: usize) -> u32

Gets a register from the virtual config space. Only used by PCI.

reg_idx - The index of the config register to read.

fn on_sandboxed(&mut self)

Invoked when the device is sandboxed.

fn get_ranges(&self) -> Vec<(BusRange, BusType)>

Gets a list of all ranges registered by this BusDevice.

fn destroy_device(&mut self)

Invoked when the device is destroyed

fn is_bridge(&self) -> Option<u8>

Returns the secondary bus number if this bus device is pci bridge