Module devices::irqchip::apic

Structs

Apic

DestinationFormat

Interrupt

An interrupt to be sent to one or more Apics.

InterruptCommand

InterruptData

The vector and type of an Interrupt.

InterruptDestination

Specifies the destination processors for an Interrupt.

LocalInterrupt

LogicalDestination

PendingInterrupts

Pending Apic interrupts to be injected into a vcpu.

Reg

VersionRegister

Enums

ApicBusMsg

A message from an Apic to the UserspaceIrqChip.

DestinationShorthand

A quick method of specifying all processors, all excluding self, or self as the destination.

VectorReg

The APIC registers that store interrupt vector bitmaps. Each has 256 bit flags, one for each interrupt vector. The flags are spread across the first 32 bits of each of eight 16-byte APIC register slots.

Constants

APIC_BASE_ADDRESS

Address of the start of APIC MMIO region.

APIC_MEM_LENGTH_BYTES

Length in bytes of APIC MMIO region.

BOOTSTRAP_PROCESSOR

CYCLE_LENGTH_FALLBACK

DESTINATION_FORMAT_CLUSTER

DESTINATION_FORMAT_FLAT

LOCAL_VECTOR_MASKED

MAX_LVT

PHYSICAL_BROADCAST_ADDRESS

REG_ALIGN_BYTES

SOFTWARE_ENABLE

TIMER_DIVIDE_TABLE

TIMER_MODE_MASK

TIMER_MODE_ONE_SHOT

TIMER_MODE_PERIODIC

TIMER_MODE_TSC_DEADLINE

VERSION

ZERO_DURATION

Type Definitions

Vector