Struct devices::irqchip::kvm::x86_64::KvmSplitIrqChip

pub struct KvmSplitIrqChip {
    vm: KvmVm,
    vcpus: Arc<Mutex<Vec<Option<KvmVcpu>>>>,
    routes: Arc<Mutex<Vec<IrqRoute>>>,
    pit: Arc<Mutex<Pit>>,
    pic: Arc<Mutex<Pic>>,
    ioapic: Arc<Mutex<Ioapic>>,
    ioapic_pins: usize,
    delayed_ioapic_irq_events: Arc<Mutex<Vec<usize>>>,
    delayed_ioapic_irq_trigger: Event,
    irq_events: Arc<Mutex<Vec<Option<IrqEvent>>>>,
}

The KvmSplitIrqsChip supports KVM’s SPLIT_IRQCHIP feature, where the PIC and IOAPIC are emulated in userspace, while the local APICs are emulated in the kernel. The SPLIT_IRQCHIP feature only supports x86/x86_64 so we only define this IrqChip in crosvm for x86/x86_64.

Fields
vm: KvmVmvcpus: Arc<Mutex<Vec<Option<KvmVcpu>>>>routes: Arc<Mutex<Vec<IrqRoute>>>pit: Arc<Mutex<Pit>>pic: Arc<Mutex<Pic>>ioapic: Arc<Mutex<Ioapic>>ioapic_pins: usizedelayed_ioapic_irq_events: Arc<Mutex<Vec<usize>>>

Vec of ioapic irq events that have been delayed because the ioapic was locked when service_irq was called on the irqchip. This prevents deadlocks when a Vcpu thread has locked the ioapic and the ioapic sends a AddMsiRoute signal to the main thread (which itself may be busy trying to call service_irq).

delayed_ioapic_irq_trigger: Event

Event which is meant to trigger process of any irqs events that were delayed.

irq_events: Arc<Mutex<Vec<Option<IrqEvent>>>>

Array of Events that devices will use to assert ioapic pins.

Implementations

impl KvmSplitIrqChip

pub fn new(
vm: KvmVm,
num_vcpus: usize,
irq_tube: Tube,
ioapic_pins: Option<usize>
) -> Result<Self>

Construct a new KvmSplitIrqChip.

impl KvmSplitIrqChip

fn routes_to_chips(&self, irq: u32) -> Vec<(IrqSourceChip, u32)>

Convenience function for determining which chips the supplied irq routes to.

fn interrupt_requested(&self, vcpu_id: usize) -> bool

Return true if there is a pending interrupt for the specified vcpu. For KvmSplitIrqChip this calls interrupt_requested on the pic.

fn get_external_interrupt(&self, vcpu_id: usize) -> Option<u32>

Check if the specified vcpu has any pending interrupts. Returns None for no interrupts, otherwise Some(u32) should be the injected interrupt vector. For KvmSplitIrqChip this calls get_external_interrupt on the pic.

fn register_irq_event(
&mut self,
irq: u32,
irq_event: &Event,
resample_event: Option<&Event>,
source: IrqEventSource
) -> Result<Option<IrqEventIndex>>

Register an event that can trigger an interrupt for a particular GSI.

fn unregister_irq_event(&mut self, irq: u32, irq_event: &Event) -> Result<()>

Unregister an event for a particular GSI.