Module devices::pci::ac97

Modules

sys

Structs

Ac97Dev

Ac97Parameters

Holds the parameters for a AC97 device

Enums

Ac97Backend

AC97 audio device emulation. Provides the PCI interface for the internal Ac97 emulation. Internally the Ac97BusMaster and Ac97Mixer structs are used to emulated the bus master and mixer registers respectively. Ac97BusMaster handles moving smaples between guest memory and the audio backend.

Ac97Error

Errors that are possible from a Ac97.

Constants

PCI_DEVICE_ID_INTEL_82801AA_5