Struct devices::pci::ac97_bus_master::Ac97BusMaster

pub struct Ac97BusMaster {
    mem: GuestMemory,
    regs: Arc<Mutex<Ac97BusMasterRegs>>,
    acc_sema: u8,
    po_info: AudioThreadInfo,
    pi_info: AudioThreadInfo,
    pmic_info: AudioThreadInfo,
    audio_server: Box<dyn ShmStreamSource<Error>>,
    irq_resample_thread: Option<JoinHandle<()>>,
}

Ac97BusMaster emulates the bus master portion of AC97. It exposes a register read/write interface compliant with the ICH bus master.

Fields
mem: GuestMemoryregs: Arc<Mutex<Ac97BusMasterRegs>>acc_sema: u8po_info: AudioThreadInfopi_info: AudioThreadInfopmic_info: AudioThreadInfoaudio_server: Box<dyn ShmStreamSource<Error>>irq_resample_thread: Option<JoinHandle<()>>
Implementations

impl Ac97BusMaster

pub(crate) fn new(
mem: GuestMemory,
audio_server: Box<dyn ShmStreamSource<Error>>
) -> Self

Creates an Ac97BusMasterobject that plays audio frommemto streams provided byaudio_server`.

pub fn keep_rds(&self) -> Option<Vec<RawDescriptor>>

Returns any file descriptors that need to be kept open when entering a jail.

pub(in pci::ac97_bus_master) fn check_and_move_to_next_buffer(
func_regs: &mut Ac97FunctionRegs
)

pub(in pci::ac97_bus_master) fn thread_semaphore_notify(
&self,
func: Ac97Function
)

fn stream_effects(func: Ac97Function) -> Vec<StreamEffect>

fn create_audio_worker(
&mut self,
mixer: &Ac97Mixer,
func: Ac97Function
) -> Result<AudioWorker, AudioError>

pub(in pci::ac97_bus_master) fn start_audio(
&mut self,
func: Ac97Function,
mixer: &Ac97Mixer
) -> Result<(), AudioError>

impl Ac97BusMaster

pub fn set_irq_event(&mut self, irq_evt: IrqLevelEvent)

Provides the events needed to raise interrupts in the guest.

pub fn update_mixer_settings(&mut self, mixer: &Ac97Mixer)

Called when mixer has been changed and the new values should be applied to currently active streams.

pub fn is_cold_reset(&self) -> bool

Checks if the bus master is in the cold reset state.

pub fn readb(&mut self, offset: u64) -> u8

Reads a byte from the given offset.

pub fn readw(&mut self, offset: u64, mixer: &Ac97Mixer) -> u16

Reads a word from the given offset.

pub fn readl(&mut self, offset: u64) -> u32

Reads a 32-bit word from the given offset.

pub fn writeb(&mut self, offset: u64, val: u8, mixer: &Ac97Mixer)

Writes the byte val to the register specified by offset.

pub fn writew(&mut self, offset: u64, val: u16)

Writes the word val to the register specified by offset.

pub fn writel(&mut self, offset: u64, val: u32, mixer: &mut Ac97Mixer)

Writes the 32-bit val to the register specified by offset.

fn set_bdbar(&mut self, func: Ac97Function, val: u32)

fn set_lvi(&mut self, func: Ac97Function, val: u8)

fn set_sr(&mut self, func: Ac97Function, val: u16)

fn set_cr(&mut self, func: Ac97Function, val: u8, mixer: &Ac97Mixer)

fn set_glob_cnt(&mut self, new_glob_cnt: u32, mixer: &mut Ac97Mixer)

fn current_sample_rate(&self, func: Ac97Function, mixer: &Ac97Mixer) -> u32

fn thread_info(&self, func: Ac97Function) -> &AudioThreadInfo

fn thread_info_mut(&mut self, func: Ac97Function) -> &mut AudioThreadInfo

fn is_audio_running(&self) -> bool

fn stop_all_audio(&mut self)

fn reset_func_regs(regs: &mut Ac97BusMasterRegs, func: Ac97Function)

fn reset_audio_regs(&mut self)