Trait devices::pci::pci_configuration::PciSubclass

pub trait PciSubclass {
    fn get_register_value(&self) -> u8;
}

A PCI sublcass. Each class in PciClassCode can specify a unique set of subclasses. This trait is implemented by each subclass. It allows use of a trait object to generate configurations.

Required methods

fn get_register_value(&self) -> u8

Convert this subclass to the value used in the PCI specification.