Trait devices::pci::pcie::pcie_device::PcieDevice

pub trait PcieDevice: Send {

    fn get_device_id(&self) -> u16;

    fn debug_label(&self) -> String;

    fn allocate_address(
        &mut self, 
        resources: &mut SystemAllocator
    ) -> Result<PciAddress, PciDeviceError>;

    fn read_config(&self, reg_idx: usize, data: &mut u32);

    fn write_config(&mut self, reg_idx: usize, offset: u64, data: &[u8]);

    fn clone_interrupt(&mut self, msi_config: Arc<Mutex<MsiConfig>>);

    fn get_caps(&self) -> Vec<Box<dyn PciCapability>>;

    fn set_capability_reg_idx(&mut self, id: PciCapabilityID, reg_idx: usize);

    fn get_removed_devices(&self) -> Vec<PciAddress>;

    fn hotplug_implemented(&self) -> bool;

    fn hotplugged(&self) -> bool;

    fn get_bridge_window_size(&self) -> (u64, u64);

    fn preferred_address(&self) -> Option<PciAddress> { ... }

    fn get_bus_range(&self) -> Option<PciBridgeBusRange> { ... }

}

Required methods

fn get_device_id(&self) -> u16

fn debug_label(&self) -> String

fn allocate_address(
&mut self,
resources: &mut SystemAllocator
) -> Result<PciAddress, PciDeviceError>

fn read_config(&self, reg_idx: usize, data: &mut u32)

fn write_config(&mut self, reg_idx: usize, offset: u64, data: &[u8])

fn clone_interrupt(&mut self, msi_config: Arc<Mutex<MsiConfig>>)

fn get_caps(&self) -> Vec<Box<dyn PciCapability>>

fn set_capability_reg_idx(&mut self, id: PciCapabilityID, reg_idx: usize)

fn get_removed_devices(&self) -> Vec<PciAddress>

fn hotplug_implemented(&self) -> bool

Hotplug capability is implemented on this bridge or not. Return true, the children pci devices could be connected through hotplug Return false, the children pci devices should be connected statically

fn hotplugged(&self) -> bool

This function returns true if this pcie device is hotplugged into the system

fn get_bridge_window_size(&self) -> (u64, u64)

Get bridge window size to cover children’s mmio size (u64, u64) -> (non_prefetchable window size, prefetchable_window_size)

Provided methods

fn preferred_address(&self) -> Option<PciAddress>

fn get_bus_range(&self) -> Option<PciBridgeBusRange>