Module devices::pci::stub

Implements a stub PCI device. This can be used to put a device on the PCI bus that will show up in PCI device enumeration with the configured parameters. The device will otherwise be non-functional, in particular it doesn’t have any BARs, IRQs etc. and neither will it handle config register interactions.

The motivation for stub PCI devices is the case of multifunction PCI devices getting passed through via VFIO to the guest. Per PCI device enumeration, functions other than 0 will only be scanned if function 0 is present. A stub PCI device is useful in that situation to present something to the guest on function 0.

Structs

NumericPciProgrammingInterface

NumericPciSubClass

StubPciDevice

StubPciParameters