Struct devices::pci::vfio_pci::VfioPciDevice

pub struct VfioPciDevice {

    device: Arc<VfioDevice>,
    config: VfioPciConfig,
    hotplug: bool,
    hotplug_bus_number: Option<u8>,
    preferred_address: PciAddress,
    pci_address: Option<PciAddress>,
    interrupt_evt: Option<IrqLevelEvent>,
    mmio_regions: Vec<PciBarConfiguration>,
    io_regions: Vec<PciBarConfiguration>,
    msi_cap: Option<VfioMsiCap>,
    msix_cap: Option<Arc<Mutex<VfioMsixCap>>>,
    irq_type: Option<VfioIrqType>,
    vm_socket_mem: Tube,
    device_data: Option<DeviceData>,
    kill_evt: Option<Event>,
    worker_thread: Option<JoinHandle<VfioPciWorker>>,
    vm_socket_vm: Option<Tube>,
    sysfs_path: Option<PathBuf>,
    header_type_reg: Option<u32>,
    ext_caps: Vec<ExtCap>,
    is_intel_lpss: bool,
    i2c_devs: HashMap<u16, PathBuf>,
    mapped_mmio_bars: BTreeMap<usize, (u64, Vec<MemSlot>)>,

}

Implements the Vfio Pci device, then a pci device is added into vm

Fields
device: Arc<VfioDevice>config: VfioPciConfighotplug: boolhotplug_bus_number: Option<u8>preferred_address: PciAddresspci_address: Option<PciAddress>interrupt_evt: Option<IrqLevelEvent>mmio_regions: Vec<PciBarConfiguration>io_regions: Vec<PciBarConfiguration>msi_cap: Option<VfioMsiCap>msix_cap: Option<Arc<Mutex<VfioMsixCap>>>irq_type: Option<VfioIrqType>vm_socket_mem: Tubedevice_data: Option<DeviceData>kill_evt: Option<Event>worker_thread: Option<JoinHandle<VfioPciWorker>>vm_socket_vm: Option<Tube>sysfs_path: Option<PathBuf>header_type_reg: Option<u32>ext_caps: Vec<ExtCap>is_intel_lpss: booli2c_devs: HashMap<u16, PathBuf>mapped_mmio_bars: BTreeMap<usize, (u64, Vec<MemSlot>)>
Implementations

impl VfioPciDevice

pub fn new(
sysfs_path: &Path,
device: VfioDevice,
hotplug: bool,
hotplug_bus_number: Option<u8>,
guest_address: Option<PciAddress>,
vfio_device_socket_msi: Tube,
vfio_device_socket_msix: Tube,
vfio_device_socket_mem: Tube,
vfio_device_socket_vm: Option<Tube>,
is_intel_lpss: bool
) -> Result<Self, PciDeviceError>

Constructs a new Vfio Pci device for the give Vfio device

pub fn pci_address(&self) -> Option<PciAddress>

Gets the pci address of the device, if one has already been allocated.

fn is_intel_gfx(&self) -> bool

fn find_region(&self, addr: u64) -> Option<PciBarConfiguration>

fn enable_intx(&mut self)

fn disable_intx(&mut self)

fn disable_irqs(&mut self)

fn enable_msi(&mut self)

fn disable_msi(&mut self)

fn enable_msix(&mut self)

fn disable_msix(&mut self)

fn msix_vectors_update(&self) -> Result<(), VfioError>

fn msix_vector_update(&self, index: usize, irqfd: Option<&Event>)

fn adjust_bar_mmap(
&self,
bar_mmaps: Vec<vfio_region_sparse_mmap_area>,
remove_mmaps: &[AddressRange]
) -> Vec<vfio_region_sparse_mmap_area>

fn remove_bar_mmap_msix(
&self,
bar_index: u32,
bar_mmaps: Vec<vfio_region_sparse_mmap_area>
) -> Vec<vfio_region_sparse_mmap_area>

fn remove_bar_mmap_lpss(
&self,
bar_index: u32,
bar_mmaps: Vec<vfio_region_sparse_mmap_area>
) -> Vec<vfio_region_sparse_mmap_area>

fn add_bar_mmap(&self, index: u32, bar_addr: u64) -> Vec<MemSlot>

fn remove_bar_mmap(&self, mmap_slots: &[MemSlot])

fn disable_bars_mmap(&mut self)

fn commit_bars_mmap(&mut self)

fn close(&mut self)

fn start_work_thread(&mut self)

fn collect_bars(&mut self) -> Vec<PciBarConfiguration>

fn configure_barmem(&mut self, bar_info: &PciBarConfiguration, bar_addr: u64)

fn allocate_root_barmem(
&mut self,
mem_bars: &[PciBarConfiguration],
resources: &mut SystemAllocator
) -> Result<Vec<BarRange>, PciDeviceError>

fn allocate_nonroot_barmem(
&mut self,
mem_bars: &mut [PciBarConfiguration],
resources: &mut SystemAllocator
) -> Result<Vec<BarRange>, PciDeviceError>

pub fn get_max_iova(&self) -> u64

Return the supported iova max address of the Vfio Pci device

fn coordinated_pm(sysfs_path: &Path, enter: bool) -> Result<()>

fn coordinated_pm_i2c_adap(
adap_path: &Path,
i2c_devs: &mut HashMap<u16, PathBuf>
) -> Result<()>

fn coordinated_pm_i2c_platdev(
plat_path: &Path,
i2c_devs: &mut HashMap<u16, PathBuf>
) -> Result<()>

fn coordinated_pm_i2c(
sysfs_path: &Path,
i2c_devs: &mut HashMap<u16, PathBuf>
) -> Result<()>

fn power_state(&self) -> Result<u8>

fn op_call(path: &Path, id: u8) -> Result<()>

fn get_ext_cap_by_reg(&self, reg: u32) -> Option<ExtCap>

fn is_skipped_reg(&self, reg: u32) -> bool