Struct devices::serial::Serial

pub struct Serial {

    interrupt_enable: Arc<AtomicU8>,
    interrupt_identification: u8,
    interrupt_evt: Event,
    line_control: u8,
    line_status: u8,
    modem_control: u8,
    modem_status: u8,
    scratch: u8,
    baud_divisor: u16,
    in_buffer: VecDeque<u8>,
    in_channel: Option<Receiver<u8>>,
    input: Option<Box<dyn SerialInput>>,
    out: Option<Box<dyn Write + Send>>,

}

Emulates serial COM ports commonly seen on x86 I/O ports 0x3f8/0x2f8/0x3e8/0x2e8.

This can optionally write the guest’s output to a Write trait object. To send input to the guest, use queue_input_bytes directly, or give a Read trait object which will be used queue bytes when used_command is called.

Fields
interrupt_enable: Arc<AtomicU8>interrupt_identification: u8interrupt_evt: Eventline_control: u8line_status: u8modem_control: u8modem_status: u8scratch: u8baud_divisor: u16in_buffer: VecDeque<u8>in_channel: Option<Receiver<u8>>input: Option<Box<dyn SerialInput>>out: Option<Box<dyn Write + Send>>
Implementations

impl Serial

pub(in serial) fn system_handle_write(
&mut self,
v: u8
) -> Result<()>

impl Serial

fn new_common(
interrupt_evt: Event,
input: Option<Box<dyn SerialInput>>,
out: Option<Box<dyn Write + Send>>
) -> Serial

pub fn device_id() -> DeviceId

Returns a unique ID for the serial device.

pub fn debug_label() -> String

Returns a debug label for the serial device. Used when setting up IrqEventSource.

pub fn queue_input_bytes(&mut self, c: &[u8]) -> Result<()>

Queues raw bytes for the guest to read and signals the interrupt if the line status would change. These bytes will be read by the guest before any bytes from the input stream that have not already been queued.

fn spawn_input_thread(&mut self)

fn handle_input_thread(&mut self)

pub fn interrupt_event(&self) -> &Event

Gets the interrupt event used to interrupt the driver when it needs to respond to this device.

fn is_dlab_set(&self) -> bool

fn is_recv_intr_enabled(&self) -> bool

fn is_thr_intr_enabled(&self) -> bool

fn is_loop(&self) -> bool

fn add_intr_bit(&mut self, bit: u8)

fn del_intr_bit(&mut self, bit: u8)

fn trigger_thr_empty(&mut self) -> Result<()>

fn trigger_recv_interrupt(&mut self) -> Result<()>

fn trigger_interrupt(&mut self) -> Result<()>

fn set_data_bit(&mut self)

fn iir_reset(&mut self)

fn handle_write(&mut self, offset: u8, v: u8) -> Result<()>