Struct devices::Ac97Dev

pub struct Ac97Dev { /* fields omitted */ }

Implementations

impl Ac97Dev

pub fn new(
mem: GuestMemory,
backend: Ac97Backend,
audio_server: Box<dyn ShmStreamSource<Error>>
) -> Self

Creates an ‘Ac97Dev’ that uses the given GuestMemory and starts with all registers at default values.

pub fn try_new(mem: GuestMemory, param: Ac97Parameters) -> Result<Self, Error>

Creates an Ac97Dev with suitable audio server inside based on Ac97Parameters. If it fails to create Ac97Dev with the given back-end, it’ll fallback to the null audio device.

pub fn minijail_policy(&self) -> &'static str

Return the minijail policy file path for the current Ac97Dev.