Trait devices::BusDeviceObj

pub trait BusDeviceObj {
    fn as_pci_device(&self) -> Option<&dyn PciDevice> { ... }

    fn as_pci_device_mut(&mut self) -> Option<&mut dyn PciDevice> { ... }

    fn into_pci_device(self: Box<Self>) -> Option<Box<dyn PciDevice>> { ... }

    fn as_platform_device(&self) -> Option<&VfioPlatformDevice> { ... }

    fn as_platform_device_mut(&mut self) -> Option<&mut VfioPlatformDevice> { ... }

    fn into_platform_device(self: Box<Self>) -> Option<Box<VfioPlatformDevice>> { ... }
}

Trait for generic device abstraction, that is, all devices that reside on BusDevice and want to be converted back to its original type. Each new foo device must provide as_foo_device() + as_foo_device_mut() + into_foo_device(), default impl methods return None.

Provided methods

fn as_pci_device(&self) -> Option<&dyn PciDevice>

fn as_pci_device_mut(&mut self) -> Option<&mut dyn PciDevice>

fn into_pci_device(self: Box<Self>) -> Option<Box<dyn PciDevice>>

fn as_platform_device(&self) -> Option<&VfioPlatformDevice>

fn as_platform_device_mut(&mut self) -> Option<&mut VfioPlatformDevice>

fn into_platform_device(self: Box<Self>) -> Option<Box<VfioPlatformDevice>>

Loading content...