Trait devices::HotPlugBus

pub trait HotPlugBus {
    fn hot_plug(&mut self, addr: PciAddress);

    fn hot_unplug(&mut self, addr: PciAddress);

    fn add_hotplug_device(
        &mut self, 
        host_key: HostHotPlugKey, 
        guest_addr: PciAddress
    );

    fn get_hotplug_device(&self, host_key: HostHotPlugKey) -> Option<PciAddress>;
}

Trait for devices that notify hotplug event into guest

Required methods

fn hot_plug(&mut self, addr: PciAddress)

Notify hotplug in event into guest

'addr' - the guest pci address for hotplug in device

fn hot_unplug(&mut self, addr: PciAddress)

Notify hotplug out event into guest

'addr' - the guest pci address for hotplug out device

fn add_hotplug_device(
&mut self,
host_key: HostHotPlugKey,
guest_addr: PciAddress
)

Add hotplug device into this bus

'host_key' - the key to identify hotplug device from host view

'guest_addr' - the guest pci address for hotplug device

fn get_hotplug_device(&self, host_key: HostHotPlugKey) -> Option<PciAddress>

get guest pci address from the specified host_key

Loading content...