Module devices::tsc

Handles operations using platform Time Stamp Counter (TSC).

Re-exports

pub use calibrate::*;

pub use cpuid::*;

Modules

calibrate

cpuid

grouping

Structs

TSC_STATE

TscSyncMitigations

Functions

get_tsc_sync_mitigations

Given the state of the host TSCs in tsc_state, and the number of vcpus that are intended to be run, return a set of affinities and TSC offsets to apply to those vcpus.

rdtsc_safe

tsc_frequency

Returns the frequency of the host TSC. Uses the TSC_STATE lazy_static ref, so the calibration only happens once.

tsc_state

Returns the state of the host TSCs. Uses the TSC_STATE lazy_static ref, so the calibration only happens once.

tsc_sync_mitigations_inner