Struct devices::usb::xhci::device_slot::DeviceSlot

pub struct DeviceSlot {
    slot_id: u8,
    port_id: PortId,
    dcbaap: Register<u64>,
    hub: Arc<UsbHub>,
    interrupter: Arc<Mutex<Interrupter>>,
    event_loop: Arc<EventLoop>,
    mem: GuestMemory,
    enabled: AtomicBool,
    transfer_ring_controllers: Mutex<Vec<Option<Arc<RingBufferController<TransferRingTrbHandler>>>>>,
}

Fields
slot_id: u8port_id: PortIddcbaap: Register<u64>hub: Arc<UsbHub>interrupter: Arc<Mutex<Interrupter>>event_loop: Arc<EventLoop>mem: GuestMemoryenabled: AtomicBooltransfer_ring_controllers: Mutex<Vec<Option<Arc<RingBufferController<TransferRingTrbHandler>>>>>
Implementations

impl DeviceSlot

pub fn new(
slot_id: u8,
dcbaap: Register<u64>,
hub: Arc<UsbHub>,
interrupter: Arc<Mutex<Interrupter>>,
event_loop: Arc<EventLoop>,
mem: GuestMemory
) -> Self

Create a new device slot.

fn get_trc(
&self,
i: usize
) -> Option<Arc<RingBufferController<TransferRingTrbHandler>>>

fn set_trc(
&self,
i: usize,
trc: Option<Arc<RingBufferController<TransferRingTrbHandler>>>
)

fn trc_len(&self) -> usize

pub fn ring_doorbell(&self, target: u8, _stream_id: u16) -> Result<bool, Error>

The arguments are identical to the fields in each doorbell register. The target value: 1: Reserved 2: Control endpoint 3: Endpoint 1 out 4: Endpoint 1 in 5: Endpoint 2 out … 32: Endpoint 15 in

Steam ID will be useful when host controller support streams. The stream ID must be zero for endpoints that do not have streams configured. This function will return false if it fails to trigger transfer ring start.

pub fn enable(&self) -> bool

Enable the slot. This function returns false if it’s already enabled.

pub fn disable<C: FnMut(TrbCompletionCode) -> Result<(), ()> + 'static + Send>(
fail_handle: Arc<dyn FailHandle>,
slot: &Arc<DeviceSlot>,
callback: C
) -> Result<(), Error>

Disable this device slot. If the slot is not enabled, callback will be invoked immediately with error. Otherwise, callback will be invoked when all trc is stopped.

pub fn set_address(
&self,
trb: &AddressDeviceCommandTrb
) -> Result<TrbCompletionCode, Error>

pub fn configure_endpoint(
&self,
trb: &ConfigureEndpointCommandTrb
) -> Result<TrbCompletionCode, Error>

pub fn evaluate_context(
&self,
trb: &EvaluateContextCommandTrb
) -> Result<TrbCompletionCode, Error>

pub fn reset_slot<C: FnMut(TrbCompletionCode) -> Result<(), ()> + 'static + Send>(
fail_handle: Arc<dyn FailHandle>,
slot: &Arc<DeviceSlot>,
callback: C
) -> Result<(), Error>

Reset the device slot to default state and deconfigures all but the control endpoint.

pub fn stop_all_trc(&self, auto_callback: RingBufferStopCallback)

Stop all transfer ring controllers.

pub fn stop_endpoint<C: FnMut(TrbCompletionCode) -> Result<(), ()> + 'static + Send>(
&self,
fail_handle: Arc<dyn FailHandle>,
endpoint_id: u8,
cb: C
) -> Result<(), Error>

Stop a endpoint.

pub fn set_tr_dequeue_ptr(
&self,
endpoint_id: u8,
ptr: u64
) -> Result<TrbCompletionCode, Error>

Set transfer ring dequeue pointer.

fn reset(&self)

fn add_one_endpoint(&self, device_context_index: u8) -> Result<(), Error>

fn drop_one_endpoint(&self, device_context_index: u8) -> Result<(), Error>

fn get_device_context(&self) -> Result<DeviceContext, Error>

fn set_device_context(&self, device_context: DeviceContext) -> Result<(), Error>

fn copy_context(
&self,
input_context_ptr: GuestAddress,
device_context_index: u8
) -> Result<(), Error>

fn get_device_context_addr(&self) -> Result<GuestAddress, Error>

fn set_state(&self, state: DeviceSlotState) -> Result<(), Error>