Module devices::usb::xhci::xhci_abi

Structs

AddressDeviceCommandTrb

AddressedTrb

POD struct associates a TRB with its address in guest memory. This is useful because transfer and command completion event TRBs must contain pointers to the original TRB that generated the event.

CommandCompletionEventTrb

ConfigureEndpointCommandTrb

DataStageTrb

DequeuePtr

DeviceContext

Device context.

DisableSlotCommandTrb

EndpointContext

EvaluateContextCommandTrb

EventDataTrb

EventRingSegmentTableEntry

InputControlContext

IsochTrb

LinkTrb

NoopTrb

NormalTrb

PortStatusChangeEventTrb

ResetDeviceCommandTrb

ResetEndpointCommandTrb

SetTRDequeuePointerCommandTrb

SetupStageTrb

SlotContext

StatusStageTrb

StopEndpointCommandTrb

TransferEventTrb

Trb

Enums

DeviceSlotState

State of device slot.

EndpointState

State of endpoint.

Error

TrbCompletionCode

Completion code of trb types.

TrbType

All kinds of trb.

Constants

DEVICE_CONTEXT_ENTRY_SIZE

SEGMENT_TABLE_SIZE

TRB_SIZE

Traits

TrbCast

All trb structs have the same size. One trb could be safely casted to another, though the values might be invalid.

TypedTrb

Associate real type of trb.

Type Definitions

Result

TransferDescriptor