Trait devices::vfio::VfioCommonTrait

pub trait VfioCommonTrait: Send + Sync {
    fn vfio_get_container(
        sysfspath: &Path, 
        iommu_enabled: bool
    ) -> Result<Arc<Mutex<VfioContainer>>, VfioError>;
}

A helper trait for managing VFIO setup

Required methods

fn vfio_get_container(
sysfspath: &Path,
iommu_enabled: bool
) -> Result<Arc<Mutex<VfioContainer>>, VfioError>

The single place to create a VFIO container for a PCI endpoint.

The policy to determine whether an individual or a shared VFIO container will be created for this device is governed by the physical PCI topology, and the argument iommu_enabled.

Arguments

sysfspath - the path to the PCI device, e.g. /sys/bus/pci/devices/0000:02:00.0

iommu_enabled - whether virtio IOMMU is enabled on this device

Loading content...