Function devices::virtio::console::handle_input

pub fn handle_input<I: SignalableInterrupt>(
    mem: &GuestMemory, 
    interrupt: &I, 
    in_channel: &Receiver<Vec<u8>>, 
    receive_queue: &mut Queue
) -> Result<(), ConsoleError>

Checks for input from in_channel_opt and transfers it to the receive queue, if any.

Arguments

mem - The GuestMemory to write the data into

interrupt - SignalableInterrupt used to signal that the queue has been used

in_channel_opt - Optional input channel to read data from

receive_queue - The receive virtio Queue