Function devices::virtio::console::spawn_input_thread

pub fn spawn_input_thread(
    rx: Box<dyn Read + Send>, 
    in_avail_evt: &Event
) -> Option<Arc<Mutex<VecDeque<u8>>>>

Starts a thread that reads rx and sends the input back via the returned buffer.

The caller should listen on in_avail_evt for events. When in_avail_evt signals that data is available, the caller should lock the returned Mutex and read data out of the inner VecDeque. The data should be removed from the beginning of the VecDeque as it is processed.

Arguments

rx - Data source that the reader thread will wait on to send data back to the buffer

in_avail_evt - Event triggered by the thread when new input is available on the buffer