Module devices::virtio::descriptor_utils

Modules

sys

Structs

DescriptorChainMemRegions

DescriptorChainRegions

Reader

Provides high-level interface over the sequence of memory regions defined by readable descriptors in the descriptor chain.

ReaderIterator

An iterator over DataInit objects on readable descriptors in the descriptor chain.

Writer

Provides high-level interface over the sequence of memory regions defined by writable descriptors in the descriptor chain.

virtq_desc

Enums

DescriptorType

Error

Constants

VIRTQ_DESC_F_NEXT

VIRTQ_DESC_F_WRITE

Functions

create_descriptor_chain

Test utility function to create a descriptor chain in guest memory.

get_mem_regions

Get all the mem regions from a DescriptorChain iterator, regardless if the DescriptorChain contains GPAs (guest physical address), or IOVAs (io virtual address). IOVAs will be translated to GPAs via IOMMU.

Type Definitions

Result