Module devices::virtio::fs::passthrough

Macros

scoped_cred

Structs

Config

Options that configure the behavior of the file system.

HandleData

InodeAltKey

InodeData

PassthroughFs

A file system that simply “passes through” all requests it receives to the underlying file system. To keep the implementation simple it servers the contents of its root directory. Users that wish to serve only a specific directory should set up the environment so that that directory ends up as the root of the file system process. One way to accomplish this is via a combination of mount namespaces and the pivot_root system call.

ScopedFsetid

ScopedGid

ScopedUid

ScopedUmask

fscrypt_get_policy_ex_arg

fscrypt_policy_v1

fscrypt_policy_v2

fsverity_digest

fsverity_enable_arg

fsxattr

Enums

CachePolicy

The caching policy that the file system should report to the FUSE client. By default the FUSE protocol uses close-to-open consistency. This means that any cached contents of the file are invalidated the next time that file is opened.

FileType

Constants

EMPTY_CSTR

FSCRYPT_KEY_DESCRIPTOR_SIZE

FSCRYPT_KEY_IDENTIFIER_SIZE

PROC_CSTR

ROOT_CSTR

SECURITY_XATTR

SELINUX_XATTR

SYS_GETEGID

SYS_GETEUID

THREAD_EGID

THREAD_EUID

USER_VIRTIOFS_XATTR

Functions

FS_IOC32_GETFLAGS

FS_IOC32_SETFLAGS

FS_IOC64_GETFLAGS

FS_IOC64_SETFLAGS

FS_IOC_ENABLE_VERITY

FS_IOC_FSGETXATTR

FS_IOC_FSSETXATTR

FS_IOC_GETFLAGS

FS_IOC_GET_ENCRYPTION_POLICY_EX

FS_IOC_MEASURE_VERITY

FS_IOC_SETFLAGS

drop_cap_fsetid

ebadf

forget_one

raise_cap_fsetid

set_creds

stat

statat

strip_xattr_prefix

Type Definitions

Handle

Inode

Unions

fscrypt_policy