Struct devices::virtio::gpu::EventDevice

pub struct EventDevice { /* fields omitted */ }

Encapsulates a virtual event device, such as a mouse or keyboard

Implementations

impl EventDevice

pub fn new(kind: EventDeviceKind, event_socket: UnixStream) -> EventDevice

pub fn mouse(event_socket: UnixStream) -> EventDevice

pub fn touchscreen(event_socket: UnixStream) -> EventDevice

pub fn keyboard(event_socket: UnixStream) -> EventDevice

pub fn kind(&self) -> EventDeviceKind

pub fn flush_buffered_events(&mut self) -> Result<bool, Error>

Flushes the buffered events that did not fit into the underlying transport, if any.

Returns Ok(true) if, after this function returns, there all the buffer of events is empty.

pub fn is_buffered_events_empty(&self) -> bool

pub fn send_report<E>(&mut self, events: E) -> Result<bool, Error> where
E: IntoIterator<Item = virtio_input_event>,
<E as IntoIterator>::IntoIter: ExactSizeIterator,

pub fn send_event_encoded(
&mut self,
event: virtio_input_event
) -> Result<bool, Error>

Sends the given event, returning Ok(true) if, after this function returns, there are no buffered events remaining.

pub fn recv_event_encoded(&self) -> Result<virtio_input_event, Error>