Module devices::virtio::input

Modules

constants

defaults

evdev

event_source

Structs

Input

Virtio input device

VirtioInputConfig

Worker

virtio_input_absinfo

virtio_input_bitmap

virtio_input_config

virtio_input_device_ids

Enums

InputError

Constants

EVENT_QUEUE_SIZE

QUEUE_SIZES

STATUS_QUEUE_SIZE

Functions

new_evdev

Creates a new virtio input device from an event device node

new_keyboard

Creates a new virtio keyboard, which supports the same events as an en-us physical keyboard.

new_mouse

Creates a new virtio mouse which supports primary, secondary, wheel and REL events.

new_multi_touch

Creates a new virtio touch device which supports multi touch.

new_single_touch

Creates a new virtio touch device which supports single touch only.

new_switches

Creates a new virtio device for switches.

new_trackpad

Creates a new virtio trackpad device which supports (single) touch, primary and secondary buttons as well as X and Y axis.

Type Definitions

Result