Module devices::virtio::input::defaults

Functions

default_keyboard_events

default_mouse_events

default_multitouchscreen_absinfo

default_multitouchscreen_events

default_switch_events

default_touchscreen_absinfo

default_touchscreen_events

default_trackpad_absinfo

default_trackpad_events

name_with_index

new_keyboard_config

Instantiates a VirtioInputConfig object with the default configuration for a keyboard. It supports the same keys as a en-us keyboard and the CAPSLOCK, NUMLOCK and SCROLLLOCK leds.

new_mouse_config

Instantiates a VirtioInputConfig object with the default configuration for a mouse. It supports left, right and middle buttons, as wel as X, Y and wheel relative axes.

new_multi_touch_config

Instantiates a VirtioInputConfig object with the default configuration for a multitouch touchscreen.

new_single_touch_config

Instantiates a VirtioInputConfig object with the default configuration for a touchscreen (no multitouch support).

new_switches_config

Instantiates a VirtioInputConfig object with the default configuration for a collection of switches.

new_trackpad_config

Instantiates a VirtioInputConfig object with the default configuration for a trackpad. It supports touch, left button and right button events, as well as X and Y axis.