Module devices::virtio::input::evdev

Structs

evdev_abs_info

evdev_buffer

evdev_id

Constants

EVDEV

Functions

EVIOCGABS

EVIOCGBIT

EVIOCGID

EVIOCGNAME

EVIOCGPROP

EVIOCGRAB

EVIOCGUNIQ

abs_info

Gets the absolute axes of an event device (see EVIOCGABS ioctl for details).

device_ids

Gets id information from an event device (see EVIOCGID ioctl for details).

errno

grab_evdev

Grabs an event device (see EVIOCGGRAB ioctl for details). After this function succeeds the given descriptor has exclusive access to the device, effectively making it unusable for any other process in the host.

name

Gets the name of an event device (see EVIOCGNAME ioctl for details).

properties

Gets the properties of an event device (see EVIOCGPROP ioctl for details).

serial_name

Gets the unique (serial) name of an event device (see EVIOCGUNIQ ioctl for details).

supported_events

Gets the event types supported by an event device as well as the event codes supported for each type (see EVIOCGBIT ioctl for details).

ungrab_evdev