Module devices::virtio::iommu::ipc_memory_mapper

Provide utility to communicate with an iommu in another process

Structs

CreateIpcMapperRet

ExportedRegion

A region exported from the virtio-iommu.

ExportedRegionInner

IpcMemoryMapper

Sends an addr translation request to another process using Tube, and gets the translated addr from another Tube

Enums

IommuRequest

IommuResponse

Functions

create_ipc_mapper

Returns a new IpcMemoryMapper instance and a response_tx for the iommu to respond to TranslateRequests.

map_bad_resp