Module devices::virtio::queue

Macros

accessors

Structs

AvailIter

Consuming iterator over all available descriptor chain heads in the queue.

Desc

DescIter

An iterator over a single descriptor chain. Not to be confused with AvailIter, which iterates over the descriptor chain heads in a queue.

DescriptorChain

A virtio descriptor chain.

Queue

A virtio queue’s parameters.

Constants

VIRTQ_AVAIL_F_NO_INTERRUPT

VIRTQ_DESC_F_INDIRECT

VIRTQ_DESC_F_NEXT

VIRTQ_DESC_F_WRITE

VIRTQ_USED_F_NO_NOTIFY