Module devices::virtio::snd::common_backend::async_funcs

Functions

defer_pcm_response_to_worker

drain_desc_receiver

handle_ctrl_queue

Handle all the control messages from the ctrl queue.

handle_event_queue

Send events to the audio driver.

handle_pcm_queue

Handle messages from the tx or the rx queue. One invocation is needed for each queue.

process_pcm_ctrl

read_data

send_pcm_response_with_writer

send_pcm_response_worker

start_pcm_worker

Start a pcm worker that receives descriptors containing PCM frames (audio data) from the tx/rx queue, and forward them to CRAS. One pcm worker per stream.

write_data