Function devices::virtio::snd::vios_backend::shm_streams::mmap_buffer

fn mmap_buffer(
    src: &mut SharedMemory, 
    offset: usize, 
    size: usize
) -> Result<(MemoryMapping, usize), Error>

Memory map a shared memory object to access an audio buffer. The buffer may not be located at an offset aligned to page size, so the offset within the mapped region is returned along with the MemoryMapping struct.