Trait devices::virtio::SignalableInterrupt

pub trait SignalableInterrupt {
    fn signal(&self, vector: u16, interrupt_status_mask: u32);

    fn signal_config_changed(&self);

    fn get_resample_evt(&self) -> Option<&Event>;

    fn do_interrupt_resample(&self);

    fn signal_used_queue(&self, vector: u16) { ... }
}

Required methods

fn signal(&self, vector: u16, interrupt_status_mask: u32)

Writes to the irqfd to VMM to deliver virtual interrupt to the guest.

fn signal_config_changed(&self)

Notify the driver that the device configuration has changed.

fn get_resample_evt(&self) -> Option<&Event>

Get the event to signal resampling is needed if it exists.

fn do_interrupt_resample(&self)

Reads the status and writes to the interrupt event. Doesn't read the resample event, it assumes the resample has been requested.

Loading content...
Provided methods

fn signal_used_queue(&self, vector: u16)

Notify the driver that buffers have been placed in the used queue.

Loading content...