Module devices::virtio::vhost::user::device::handler

Library for implementing vhost-user device executables.

This crate provides

VhostUserBackend trait, which is a collection of methods to handle vhost-user requests, and

DeviceRequestHandler struct, which makes a connection to a VMM and starts an event loop.

They are expected to be used as follows:

Define a struct and implement VhostUserBackend for it.

Create a DeviceRequestHandler with the backend struct.

Drive the DeviceRequestHandler::run async fn with an executor.

struct MyBackend {
  /* fields */
}

impl VhostUserBackend for MyBackend {
  /* implement methods */
}

fn main() -> Result<(), Box<dyn Error>> {
  let backend = MyBackend { /* initialize fields */ };
  let handler = DeviceRequestHandler::new(backend);
  let socket = std::path::Path("/path/to/socket");
  let ex = cros_async::Executor::new()?;

  if let Err(e) = ex.run_until(handler.run(socket, &ex)) {
    eprintln!("error happened: {}", e);
  }
  Ok(())
}

Modules

sys

Structs

CallEvent

An event to deliver an interrupt to the guest.

DeviceRequestHandler

Structure to have an event loop for interaction between a VMM and VhostUserBackend.

MappingInfo

Keeps a mapping from the vmm’s virtual addresses to guest addresses. used to translate messages from the vmm to guest offsets.

VhostShmemMapper

VhostUserRegularOps

Ops for running vhost-user over a stream (i.e. regular protocol).

Vring

A virtio ring entry.

Traits

VhostUserBackend

Trait for vhost-user backend.

VhostUserPlatformOps

Trait for defining vhost-user ops that are platform-dependent.

Functions

vmm_va_to_gpa