Trait devices::virtio::vhost::user::device::handler::sys::unix::SharedMemorySys

pub(crate) trait SharedMemorySys {
    fn get_seals(&self) -> Result<MemfdSeals, Error>;

    fn add_seals(&mut self, seals: MemfdSeals) -> Result<(), Error>;

    fn from_safe_descriptor(
        descriptor: SafeDescriptor
    ) -> Result<SharedMemory, Error> { ... }

    fn from_file(file: File) -> Result<SharedMemory, Error> { ... }
}

Required methods

fn get_seals(&self) -> Result<MemfdSeals, Error>

fn add_seals(&mut self, seals: MemfdSeals) -> Result<(), Error>

Provided methods

fn from_safe_descriptor(
descriptor: SafeDescriptor
) -> Result<SharedMemory, Error>

Creates a SharedMemory instance from a SafeDescriptor owning a reference to a shared memory descriptor. Ownership of the underlying descriptor is transferred to the new SharedMemory object.

fn from_file(file: File) -> Result<SharedMemory, Error>