Trait devices::virtio::vhost::user::device::handler::VhostUserPlatformOps

pub trait VhostUserPlatformOps {
    fn protocol(&self) -> Protocol;

    fn set_mem_table(
        &mut self, 
        contexts: &[VhostUserMemoryRegion], 
        files: Vec<File>
    ) -> VhostResult<(GuestMemory, Vec<MappingInfo>)>;

    fn set_vring_kick(
        &mut self, 
        index: u8, 
        file: Option<File>
    ) -> VhostResult<Event>;

    fn set_vring_call(
        &mut self, 
        index: u8, 
        file: Option<File>
    ) -> VhostResult<Doorbell>;
}

Trait for defining vhost-user ops that are platform-dependent.

Required methods

fn protocol(&self) -> Protocol

Returns the protocol implemented by these platform ops.

fn set_mem_table(
&mut self,
contexts: &[VhostUserMemoryRegion],
files: Vec<File>
) -> VhostResult<(GuestMemory, Vec<MappingInfo>)>

Create the guest memory for the backend.

contexts and files must be the same size, and provide a description of the memory regions to map as well as the file descriptors from which to obtain the memory backing these regions, respectively.

The returned tuple contains the constructed GuestMemory from these memory contexts, as well as a vector describing all the mappings described by these contexts.

fn set_vring_kick(
&mut self,
index: u8,
file: Option<File>
) -> VhostResult<Event>

Return an Event that will be signaled by the frontend whenever vring index should be processed.

For protocols that support providing that event using a file descriptor (Regular), it is provided by file. For other protocols, file will be None.

fn set_vring_call(
&mut self,
index: u8,
file: Option<File>
) -> VhostResult<Doorbell>

Return a Doorbell that the backend will signal whenever it puts used buffers for vring index.

For protocols that support listening to a file descriptor (Regular), file provides a file descriptor from which the Doorbell should be built. For other protocols, it will be None.