Trait devices::virtio::vhost::user::device::VhostUserDevice

pub trait VhostUserDevice {
    fn max_queue_num(&self) -> usize;

    fn into_backend(
        self: Box<Self>, 
        ex: &Executor
    ) -> Result<Box<dyn VhostUserBackend>>;
}

A trait for vhost-user devices.

Upon being given an [Executor], a device can be converted into a [VhostUserBackend], which can then process the requests from the front-end.

We don’t build VhostUserBackends directly because in the case of jailing, the device is built in the main process but it runs in the jailed child process. Since Executors cannot be passed to other processes, we cannot access the device’s executor at build time and thus need to perform this 2-step dance before we can run the vhost-user device jailed.

Required methods

fn max_queue_num(&self) -> usize

The maximum number of queues that this device can manage.

fn into_backend(
self: Box<Self>,
ex: &Executor
) -> Result<Box<dyn VhostUserBackend>>

Turn this device into a VhostUserBackend, ready to process requests.

If the device needs to perform something after being jailed, this is also the right place to do it.