Trait devices::virtio::vhost::user::device::vvu::queue::IovaAllocator

pub trait IovaAllocator {
    fn alloc_iova(&self, size: u64, tag: u8) -> Result<u64>;

    unsafe fn map_iova(
        &self, 
        iova: u64, 
        size: u64, 
        addr: *const u8
    ) -> Result<()>;
}

Interface used by UserQueue to interact with the IOMMU.

Required methods

fn alloc_iova(&self, size: u64, tag: u8) -> Result<u64>

Allocates an IO virtual address region of the requested size.

unsafe fn map_iova(&self, iova: u64, size: u64, addr: *const u8) -> Result<()>

Maps the given address at the given IOVA.

Safety

addr must reference a region of at least length size. Memory passed to this function may be mutated at any time, so addr must not be memory that is directly managed by rust.