Enum devices::virtio::vhost::user::proxy::Error

pub enum Error {

    AcceptConnection(Error),
    BarNotAllocated(usize),
    CallEventNotSet(usize),
    CreateListener(Error),
    CreateWaitContext(Error),
    CreateWriter,
    FailedAck(VhostError),
    FailedToAcceptSiblingConnection(Error),
    FailedToReadKickEvt(Error, usize),
    FailedToReceiveDoorbellData(TubeError),
    FailedToWriteCallEvent(Error, usize),
    InvalidBar(usize),
    InvalidSiblingMessage,
    KickDataNotSet(usize),
    MemoryMappingRequestFailure,
    MsiVectorNotSet(usize),
    ParseVringFdRequest(VhostError),
    ReadSiblingHeader(VhostError),
    ReadSiblingPayload(VhostError),
    RxBufferTooSmall,
    RxDescriptorsExhausted,
    SiblingDisconnected,
    TubeReceiveFailure(TubeError),
    TubeSendFailure(TubeError),
    WaitContextAddKickEvent(Error),
    WaitContextDisableSiblingVmSocket(Error),
    WaitContextEnableSiblingVmSocket(Error),
    WaitError(Error),
    WriteBuffer(Error),
    WriterCreation(DescriptorUtilsError),

}

Variants

AcceptConnection(Error)

Tuple Fields

0: Error

Failed to accept connection on a socket.

BarNotAllocated(usize)

Tuple Fields

0: usize

Bar not allocated.

CallEventNotSet(usize)

Tuple Fields

0: usize

Call event not set for a vring.

CreateListener(Error)

Tuple Fields

0: Error

Failed to create a listener.

CreateWaitContext(Error)

Tuple Fields

0: Error

Failed to create a wait context object.

CreateWriter

Failed to create a Writer object.

FailedAck(VhostError)

Tuple Fields

0: VhostError

Failed to send ACK in response to Vhost-user sibling message.

FailedToAcceptSiblingConnection(Error)

Tuple Fields

0: Error

Failed to accept sibling connection.

FailedToReadKickEvt(Error, usize)

Tuple Fields

0: Error

1: usize

Failed to read kick event for a vring.

FailedToReceiveDoorbellData(TubeError)

Tuple Fields

0: TubeError

Failed to receive doorbell data.

FailedToWriteCallEvent(Error, usize)

Tuple Fields

0: Error

1: usize

Failed to write call event.

InvalidBar(usize)

Tuple Fields

0: usize

Invalid PCI bar index.

InvalidSiblingMessage

Invalid Vhost-user sibling message.

KickDataNotSet(usize)

Tuple Fields

0: usize

Kick data not set for a vring.

MemoryMappingRequestFailure

Failed to send a memory mapping request.

MsiVectorNotSet(usize)

Tuple Fields

0: usize

MSI vector not set for a vring.

ParseVringFdRequest(VhostError)

Tuple Fields

0: VhostError

Failed to parse vring kick / call file descriptors.

ReadSiblingHeader(VhostError)

Tuple Fields

0: VhostError

Failed to read payload of a Vhost-user sibling header.

ReadSiblingPayload(VhostError)

Tuple Fields

0: VhostError

Failed to read payload of a Vhost-user sibling message.

RxBufferTooSmall

Rx buffer too small to accomodate data.

RxDescriptorsExhausted

There are no more available descriptors to receive into.

SiblingDisconnected

Sibling is disconnected.

TubeReceiveFailure(TubeError)

Tuple Fields

0: TubeError

Failed to receive a memory mapping request from the main process.

TubeSendFailure(TubeError)

Tuple Fields

0: TubeError

Failed to send a memory mapping request to the main process.

WaitContextAddKickEvent(Error)

Tuple Fields

0: Error

Adding |SET_VRING_KICK| related epoll event failed.

WaitContextDisableSiblingVmSocket(Error)

Tuple Fields

0: Error

Removing read event from the sibling VM socket events failed.

WaitContextEnableSiblingVmSocket(Error)

Tuple Fields

0: Error

Adding read event to the sibling VM socket events failed.

WaitError(Error)

Tuple Fields

0: Error

Failed to wait for events.

WriteBuffer(Error)

Tuple Fields

0: Error

Writing to a buffer in the guest failed.

WriterCreation(DescriptorUtilsError)

Tuple Fields

0: DescriptorUtilsError

Failed to create a Writer.