Enum devices::virtio::vhost::user::vmm::Error

pub enum Error {
    CopyConfig(Error),
    CreateEvent(Error),
    GetConfig(VhostError),
    GetFeatures(VhostError),
    GetHostAddress(GuestMemoryError),
    GetProtocolFeatures(VhostError),
    GetQueueNum(VhostError),
    GetVringBase(VhostError),
    InvalidConfigOffset {
        data_len: u64,
        offset: u64,
        config_len: u64,
    },
    MsixConfigUnavailable,
    MsixIrqfdUnavailable,
    ResetOwner(VhostError),
    SetConfig(VhostError),
    SetDeviceRequestChannel(VhostError),
    SetFeatures(VhostError),
    SetMemTable(VhostError),
    SetOwner(VhostError),
    SetProtocolFeatures(VhostError),
    SetVringAddr(VhostError),
    SetVringBase(VhostError),
    SetVringCall(VhostError),
    SetVringEnable(VhostError),
    SetVringKick(VhostError),
    SetVringNum(VhostError),
    SocketConnect(Error),
    SocketConnectOnMasterCreate(VhostError),
    TagTooLong {
        len: usize,
        max: usize,
    },
}

Variants

CopyConfig(Error)

Failed to copy config to a buffer.

CreateEvent(Error)

Failed to create base::Event.

GetConfig(VhostError)

Failed to get config.

GetFeatures(VhostError)

Failed to get features.

GetHostAddress(GuestMemoryError)

Failed to get host address.

GetProtocolFeatures(VhostError)

Failed to get protocol features.

GetQueueNum(VhostError)

Failed to get number of queues.

GetVringBase(VhostError)

Failed to get vring base offset.

InvalidConfigOffset

Invalid config offset is given.

Fields of InvalidConfigOffset

data_len: u64offset: u64config_len: u64

MsixConfigUnavailable

MSI-X config is unavailable.

MsixIrqfdUnavailable

MSI-X irqfd is unavailable.

ResetOwner(VhostError)

Failed to reset owner.

SetConfig(VhostError)

Failed to set config.

SetDeviceRequestChannel(VhostError)

Failed to set device request channel.

SetFeatures(VhostError)

Failed to set features.

SetMemTable(VhostError)

Failed to set memory map regions.

SetOwner(VhostError)

Failed to set owner.

SetProtocolFeatures(VhostError)

Failed to set protocol features.

SetVringAddr(VhostError)

Failed to set vring address.

SetVringBase(VhostError)

Failed to set vring base offset.

SetVringCall(VhostError)

Failed to set eventfd to signal used vring buffers.

SetVringEnable(VhostError)

Failed to enable or disable vring.

SetVringKick(VhostError)

Failed to set eventfd for adding buffers to vring.

SetVringNum(VhostError)

Failed to set the size of the queue.

SocketConnect(Error)

Failed to connect socket.

SocketConnectOnMasterCreate(VhostError)

Failed to create Master from a UDS path.

TagTooLong

The tag for the Fs device was too long to fit in the config space.

Fields of TagTooLong

len: usizemax: usize