Struct devices::virtio::vhost::user::vmm::VhostUserHandler

pub struct VhostUserHandler {
    pub avail_features: u64,
    // some fields omitted
}

Fields
avail_features: u64
Implementations

impl VhostUserHandler

pub fn new_from_path<P: AsRef<Path>>(
path: P,
max_queue_num: u64,
allow_features: u64,
init_features: u64,
allow_protocol_features: VhostUserProtocolFeatures
) -> Result<Self>

Creates a VhostUserHandler instance attached to the provided UDS path with features and protocol features initialized.

pub fn new_from_stream(
sock: UnixStream,
max_queue_num: u64,
allow_features: u64,
init_features: u64,
allow_protocol_features: VhostUserProtocolFeatures
) -> Result<Self>

Creates a VhostUserHandler instance attached to the provided UnixStream with features and protocol features initialized.

pub fn queue_sizes(
&mut self,
queue_size: u16,
default_queues_num: usize
) -> Result<Vec<u16>>

Returns a vector of sizes of each queue.

pub fn ack_features(&mut self, ack_features: u64) -> Result<()>

Enables a set of features.

pub fn read_config<T>(&mut self, offset: u64, data: &mut [u8]) -> Result<()>

Gets the device configuration space at offset and writes it into data.

pub fn write_config<T>(&mut self, offset: u64, data: &[u8]) -> Result<()>

Writes data into the device configuration space at offset.

pub fn set_device_request_channel(&mut self, channel: Tube) -> Result<()>

Sets the channel for device-specific messages.

pub fn set_mem_table(&mut self, mem: &GuestMemory) -> Result<()>

Sets the memory map regions so it can translate the vring addresses.

pub fn activate_vring(
&mut self,
mem: &GuestMemory,
queue_index: usize,
queue: &Queue,
queue_evt: &Event,
irqfd: &Event
) -> Result<()>

Activates a vring for the given queue.

pub fn activate(
&mut self,
mem: &GuestMemory,
interrupt: &Interrupt,
queues: &[Queue],
queue_evts: &[Event]
) -> Result<()>

Activates vrings.

pub fn reset(&mut self, queues_num: usize) -> Result<()>

Deactivates all vrings.