Module devices::virtio::video::decoder

Implementation of a virtio video decoder backed by a device.

Modules

backend

This module implements the interface that actual decoder devices need to implement in order to provide video decoding capability to the guest.

capability

Capablities of the virtio video decoder device.

Structs

Context

ContextMap

A thin wrapper of a map of contexts with error handlings.

Decoder

Represents information of a decoder backed by a DecoderBackend.

InputResources

OutputResources

Enums

PendingResponse

QueueOutputResourceResult

Type Definitions

FrameBufferId

InputResourceId

OutputResourceId

ResourceId

StreamId

Timestamp