Module devices::virtio::video::decoder::backend

This module implements the interface that actual decoder devices need to implement in order to provide video decoding capability to the guest.

Modules

ffmpeg

A ffmpeg-based software decoder backend for crosvm. Since it does not require any particular harware, it can provide fake hardware acceleration decoding to any guest and is mostly useful in order to work on the virtio-video specification, or to implement guest decoder code from the comfort of a workstation.

utils

Enums

DecoderEvent

Traits

DecoderBackend

DecoderSession

Contains the device’s state for one playback session, i.e. one stream.