Struct devices::virtio::video::decoder::backend::vaapi::VaapiDecoderSession

pub struct VaapiDecoderSession {
    codec: Box<dyn VaapiCodec>,
    output_queue_state: OutputQueueState,
    ready_queue: VecDeque<DecodedFrameHandle>,
    event_queue: EventQueue<DecoderEvent>,
    flushing: bool,
    coded_format: Format,
}

A decoder session for the libva backend

Fields
codec: Box<dyn VaapiCodec>

The implementation for the codec specific logic.

output_queue_state: OutputQueueState

The state for the output queue. Updated when set_output_buffer_count is called or when we detect a dynamic resolution change.

ready_queue: VecDeque<DecodedFrameHandle>

Queue containing decoded pictures.

event_queue: EventQueue<DecoderEvent>

The event queue we can use to signal new events.

flushing: bool

Whether the decoder is currently flushing.

coded_format: Format

The coded format we are decoding.

Implementations

impl VaapiDecoderSession

fn create_surfaces(
display: Rc<Display>,
width: u32,
height: u32,
rt_format: u32,
num_surfaces: usize
) -> Result<Vec<Surface>>

Create the num_surfaces surfaces for a combination of width, height and rt_format.

fn change_resolution(&mut self, new_params: DrcParams) -> Result<()>

fn nv12_copy(
src: &[u8],
dst: &mut [u8],
width: u32,
height: u32,
strides: [u32; 3],
offsets: [u32; 3]
)

Copies src into dst as NV12, removing any extra padding.

fn copy_image_to_output(
image: &Image<'_>,
output_buffer: &mut GuestResource
) -> Result<()>

Copy the decoded image into output_buffer, removing any extra padding. This is hardcoded to NV12 for now.

pub fn output_picture(
&mut self,
decoded_frame: DecodedFrameHandle,
resolution: Resolution,
image_format: &VAImageFormat
) -> Result<bool>

Copy raw decoded data from image into the output buffer

fn drain_ready_queue(&mut self) -> Result<()>