Trait devices::virtio::video::decoder::backend::vaapi::VaapiCodec

pub trait VaapiCodec: Downcast {
    fn decode(
        &mut self, 
        bitstream_id: i32, 
        resource: &GuestResourceHandle, 
        offset: u32, 
        bytes_used: u32
    ) -> Result<Vec<DecodedFrameHandle>>;

    fn flush(&mut self) -> Result<()>;

    fn va_image_fmt(&self) -> &Option<VAImageFormat>;

    fn set_raw_fmt(&mut self, format: Format) -> Result<()>;

    fn drc(&mut self) -> Option<DrcParams>;
}

Required methods

fn decode(
&mut self,
bitstream_id: i32,
resource: &GuestResourceHandle,
offset: u32,
bytes_used: u32
) -> Result<Vec<DecodedFrameHandle>>

Decode the compressed stream contained in [offset..offset+bytes_used] of the shared memory in resource. bitstream_id is the identifier for that part of the stream (most likely, a timestamp). Returns zero or more decoded pictures depending on the compressed stream, which might also be part of the codec’s decoded picture buffer (DPB).

fn flush(&mut self) -> Result<()>

Flush the decoder i.e. finish processing all queued decode requests and emit frames for them.

fn va_image_fmt(&self) -> &Option<VAImageFormat>

Returns the current VA image format

fn set_raw_fmt(&mut self, format: Format) -> Result<()>

Set the raw picture format.

fn drc(&mut self) -> Option<DrcParams>

Whether the codec has encountered a dynamic resolution change, i.e.: Whether there were changes in coded resolution (OUTPUT width and height), in the visible resolution (selection rectangles), in the minimum number of buffers needed for decoding or in bit-depth of the bitstream. This function will be called at every frame by the session.

Implementations

impl dyn VaapiCodec

pub fn is<__T: VaapiCodec>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: VaapiCodec>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: VaapiCodec>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: VaapiCodec>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: VaapiCodec>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.