Struct devices::virtio::video::decoder::backend::vaapi::vp8::Vp8Codec

pub struct Vp8Codec {
    metadata_state: StreamMetadataState,
    parser: Parser,
    last_picture: Option<DecodedFrameHandle>,
    golden_ref_picture: Option<DecodedFrameHandle>,
    alt_ref_picture: Option<DecodedFrameHandle>,
    image_fmt: Option<VAImageFormat>,
    resolution: Resolution,
    wait_keyframe: bool,
    open: bool,
    drc: bool,
}

Fields
metadata_state: StreamMetadataState

The metadata state. Updated whenever the decoder reads new data from the stream.

parser: Parser

A parser to extract bitstream data and build frame data in turn

last_picture: Option<DecodedFrameHandle>

The picture used as the last reference picture.

golden_ref_picture: Option<DecodedFrameHandle>

The picture used as the golden reference picture.

alt_ref_picture: Option<DecodedFrameHandle>

The picture used as the alternate reference picture.

image_fmt: Option<VAImageFormat>

The VAImageFormat we will decode into

resolution: Resolution

The current resolution

wait_keyframe: bool

Whether we are still waiting for the leading key frame. We skip all frames until then, since decoding can only start from key frames.

open: bool

Whether the initial setup is done after reading some metadata for the first time.

drc: bool

Whether the codec has identified a dynamic resolution change. Cleared when this information is relayed to the session.

Implementations

impl Vp8Codec

pub fn new(display: Rc<Display>) -> Result<Self>

Create a new codec backend for VP8.

fn clamp<T: PartialOrd>(x: T, low: T, high: T) -> T

A clamp such that min <= x <= max

fn change_resolution(&mut self, resolution: Resolution) -> Result<()>

fn open(&mut self, frame_hdr: &Header) -> Result<()>

Initialize the codec state by reading some metadata from the current frame.

fn check_stream_params(&mut self, frame_hdr: &Header) -> Result<()>

Check for new stream parameters. This might mean that the surfaces and/or context get recreated. If a new resolution is detected, the decoder will enter a DRC state and it will ask crosvm for new output buffers to write to.

fn build_iq_matrix(frame_hdr: &Header, parser: &Parser) -> Result<BufferType>

fn build_probability_table(frame_hdr: &Header) -> BufferType

fn build_pic_param(
frame_hdr: &Header,
resolution: &Resolution,
seg: &Segmentation,
adj: &MbLfAdjustments,
last: &Option<DecodedFrameHandle>,
golden: &Option<DecodedFrameHandle>,
alt: &Option<DecodedFrameHandle>
) -> Result<BufferType>

fn build_slice_param(
frame_hdr: &Header,
slice_size: usize
) -> Result<BufferType>

fn replace_reference(
reference: &mut Option<DecodedFrameHandle>,
picture: &DecodedFrameHandle
)

Replace a reference frame with picture.

fn update_references(
decoded_frame: &DecodedFrameHandle,
last_picture: &mut Option<DecodedFrameHandle>,
golden_ref_picture: &mut Option<DecodedFrameHandle>,
alt_ref_picture: &mut Option<DecodedFrameHandle>
) -> Result<()>

fn decode_frame<T: AsRef<[u8]>>(
&mut self,
frame: Frame<T>,
bitstream_id: i32
) -> Result<DecodedFrameHandle>

Decode a single frame, copying the decoded data into the output buffer as per the format in self.image_fmt.

fn handle_frame<T: AsRef<[u8]>>(
&mut self,
frame: Frame<T>,
bitstream_id: i32
) -> Result<Vec<DecodedFrameHandle>>

Handle a single frame.