Trait devices::virtio::video::device::Device

pub trait Device {
    fn process_cmd(
        &mut self, 
        cmd: VideoCmd, 
        wait_ctx: &WaitContext<Token>
    ) -> (VideoCmdResponseType, Option<(u32, Vec<VideoEvtResponseType>)>);

    fn process_event(
        &mut self, 
        desc_map: &mut AsyncCmdDescMap, 
        stream_id: u32
    ) -> Option<Vec<VideoEvtResponseType>>;
}

Required methods

fn process_cmd(
&mut self,
cmd: VideoCmd,
wait_ctx: &WaitContext<Token>
) -> (VideoCmdResponseType, Option<(u32, Vec<VideoEvtResponseType>)>)

Processes a virtio-video command. If the command expects a synchronous response, it returns a response as VideoCmdResponseType::Sync. Otherwise, it returns a name of the descriptor chain that will be used when a response is prepared. Implementations of this method is passed a WaitContext object which can be used to add or remove descriptors to wait on. It is expected that only Token::Event items would be added. When a Token::Event event arrives, process_event() will be invoked. TODO(b/149720783): Make this an async function.

fn process_event(
&mut self,
desc_map: &mut AsyncCmdDescMap,
stream_id: u32
) -> Option<Vec<VideoEvtResponseType>>

Processes an available Token::Event event and returns a list of VideoEvtResponseType responses. It returns None if an invalid event comes. For responses to be sent via command queue, the return type is VideoEvtResponseType::AsyncCmd. For responses to be sent via event queue, the return type is VideoEvtResponseType::Event. TODO(b/149720783): Make this an async function.