Module devices::virtio::video::format

Data structures that represent video format information in virtio video devices.

Structs

Crop

FormatDesc

FormatRange

FrameFormat

FramePlane

Description of the layout for a single plane.

PlaneFormat

Rect

A rectangle used to describe portions of a frame.

Enums

Bitrate

BitrateMode

Format

Level

Profile