Trait devices::virtio::video::resource::BufferHandle

pub trait BufferHandle: Sized {
    fn try_clone(&self) -> Result<Self, Error>;

    fn get_mapping(
        &self, 
        offset: usize, 
        size: usize
    ) -> Result<MemoryMappingArena, MmapError>;
}

Trait for types that can serve as video buffer backing memory.

Required methods

fn try_clone(&self) -> Result<Self, Error>

Try to clone this handle. This must only create a new reference to the same backing memory and not duplicate the buffer itself.

fn get_mapping(
&self,
offset: usize,
size: usize
) -> Result<MemoryMappingArena, MmapError>

Returns a linear mapping of [offset..offset+size] of the memory backing this buffer.