Crate devices

Emulates virtual and hardware devices.

Re-exports

pub use self::acpi::ACPIPMResource;

pub use self::bat::BatteryError;

pub use self::bat::GoldfishBattery;

pub use self::direct_io::DirectIo;

pub use self::direct_io::DirectMmio;

pub use self::direct_irq::DirectIrq;

pub use self::direct_irq::DirectIrqError;

pub use self::irqchip::*;

pub use self::pl030::Pl030;

pub use self::serial_device::Error as SerialError;

pub use self::serial_device::SerialDevice;

pub use self::serial_device::SerialHardware;

pub use self::serial_device::SerialParameters;

pub use self::serial_device::SerialType;

pub use self::usb::host_backend::host_backend_device_provider::HostBackendDeviceProvider;

pub use self::usb::xhci::xhci_controller::XhciController;

pub use self::vfio::VfioContainer;

pub use self::vfio::VfioDevice;

Modules

acpi

bat

direct_io

direct_irq

irqchip

pl030

serial_device

usb

vfio

virtio

Implements virtio devices, queues, and transport mechanisms.

Macros

register

register_array

static_register

Macro helps to build a static register.

usb_debug

Structs

Ac97Dev

Ac97Parameters

Holds the parameters for a AC97 device

Bus

A device container for routing reads and writes over some address space.

BusAccessInfo

Information about how a device was accessed.

BusRange

Holds a base and length representing the address space occupied by a BusDevice.

Cmos

A CMOS/RTC device commonly seen on x86 I/O port 0x70/0x71.

I8042Device

A i8042 PS/2 controller that emulates just enough to shutdown the machine.

PciAddress

PCI Device Address, AKA Bus:Device.Function

PciBridge

PciConfigIo

Emulates PCI configuration access mechanism #1 (I/O ports 0xcf8 and 0xcfc).

PciConfigMmio

Emulates PCI memory-mapped configuration access mechanism.

PciRoot

Emulates the PCI Root bridge.

PcieRootPort

Pit

ProxyDevice

Wraps an inner BusDevice that is run inside a child process via fork.

Serial

Emulates serial COM ports commonly seen on x86 I/O ports 0x3f8/0x2f8/0x3e8/0x2e8.

StubPciDevice

StubPciParameters

VfioPciDevice

Implements the Vfio Pci device, then a pci device is added into vm

VfioPlatformDevice

VirtioPciDevice

Implements the PCI transport for virtio devices.

Enums

Ac97Backend

AC97 audio device emulation. Provides the PCI interface for the internal Ac97 emulation. Internally the Ac97BusMaster and Ac97Mixer structs are used to emulated the bus master and mixer registers respectively. Ac97BusMaster handles moving smaples between guest memory and the audio backend.

BusError

BusType

HostHotPlugKey

The key to identify hotplug device from host view. like host sysfs path for vfio pci device, host disk file path for virtio block device

PciClassCode

Classes of PCI nodes.

PciDeviceError

PciInterruptPin

PCI has four interrupt pins A->D.

PitError

ProxyError

Errors for proxy devices.

Traits

BusDevice

Trait for devices that respond to reads or writes in an arbitrary address space.

BusDeviceObj

Trait for generic device abstraction, that is, all devices that reside on BusDevice and want to be converted back to its original type. Each new foo device must provide as_foo_device() + as_foo_device_mut() + into_foo_device(), default impl methods return None.

BusDeviceSync

BusResumeDevice

HotPlugBus

Trait for devices that notify hotplug event into guest

PciDevice