Struct devices::irqchip::userspace::UserspaceIrqChip

pub struct UserspaceIrqChip<V: VcpuX86_64> {
    vcpus: Arc<Mutex<Vec<Option<V>>>>,
    routes: Arc<Mutex<Routes>>,
    pit: Arc<Mutex<Pit>>,
    pic: Arc<Mutex<Pic>>,
    ioapic: Arc<Mutex<Ioapic>>,
    ioapic_pins: usize,
    apics: Vec<Arc<Mutex<Apic>>>,
    waiters: Vec<Arc<Waiter>>,
    timer_descriptors: Vec<Descriptor>,
    delayed_ioapic_irq_events: Arc<Mutex<DelayedIoApicIrqEvents>>,
    irq_events: Arc<Mutex<Vec<Option<IrqEvent>>>>,
    dropper: Arc<Mutex<Dropper>>,
}

An IrqChip with all interrupt devices emulated in userspace. UserspaceIrqChip works with any hypervisor, but only supports x86.

Fields
vcpus: Arc<Mutex<Vec<Option<V>>>>routes: Arc<Mutex<Routes>>pit: Arc<Mutex<Pit>>pic: Arc<Mutex<Pic>>ioapic: Arc<Mutex<Ioapic>>ioapic_pins: usizeapics: Vec<Arc<Mutex<Apic>>>waiters: Vec<Arc<Waiter>>timer_descriptors: Vec<Descriptor>delayed_ioapic_irq_events: Arc<Mutex<DelayedIoApicIrqEvents>>

Delayed ioapic irq object, that contains the delayed events because the ioapic was locked when service_irq was called on the irqchip. This prevents deadlocks when a Vcpu thread has locked the ioapic and the ioapic sends a AddMsiRoute signal to the main thread (which itself may be busy trying to call service_irq).

Note:

This lock may be locked by itself to access the DelayedIoApicIrqEvents. If accessed in conjunction with the irq_events field, that lock should be taken first to prevent deadlocks stemming from lock-ordering issues.

irq_events: Arc<Mutex<Vec<Option<IrqEvent>>>>dropper: Arc<Mutex<Dropper>>
Implementations

impl<V: VcpuX86_64 + 'static> UserspaceIrqChip<V>

pub fn new(
num_vcpus: usize,
irq_tube: Tube,
ioapic_pins: Option<usize>
) -> Result<Self>

Constructs a new UserspaceIrqChip.

pub fn new_with_clock(
num_vcpus: usize,
irq_tube: Tube,
ioapic_pins: Option<usize>,
clock: Arc<Mutex<Clock>>
) -> Result<Self>

Constructs a new UserspaceIrqChip, with a clock. Used for testing.

fn handle_msg(&self, msg: ApicBusMsg)

Handles a message from an APIC.

fn send_msi(&self, addr: u32, data: u32)

Sends a Message Signaled Interrupt to one or more APICs. MSIs are a 64-bit address and 32-bit data, but in the Intel spec we’re implementing, only the low 32 bits of the address are used.

fn send_irq_to_apic(&self, id: usize, irq: &InterruptData)

fn send_irq_to_apics(&self, irq: &Interrupt)

Sends an interrupt to one or more APICs. Used for sending MSIs and IPIs.

fn deliver_startup(&self, vcpu: &V, vector: u8) -> Result<()>

Delivers a startup IPI to vcpu.

fn is_runnable(&self, vcpu_id: usize) -> bool

Checks if the specified VCPU is in a runnable state.

impl<V: VcpuX86_64 + 'static> UserspaceIrqChip<V>

fn register_irq_event(
&mut self,
irq: u32,
irq_event: &Event,
resample_event: Option<&Event>,
source: IrqEventSource
) -> Result<Option<IrqEventIndex>>

fn unregister_irq_event(&mut self, irq: u32, irq_event: &Event) -> Result<()>