Trait devices::BusDevice

pub trait BusDevice: Send {
    fn debug_label(&self) -> String;

    fn read(&mut self, offset: BusAccessInfo, data: &mut [u8]) { ... }

    fn write(&mut self, offset: BusAccessInfo, data: &[u8]) { ... }

    fn config_register_write(
        &mut self, 
        reg_idx: usize, 
        offset: u64, 
        data: &[u8]
    ) -> ConfigWriteResult { ... }

    fn config_register_read(&self, reg_idx: usize) -> u32 { ... }

    fn on_sandboxed(&mut self) { ... }
}

Trait for devices that respond to reads or writes in an arbitrary address space.

The device does not care where it exists in address space as each method is only given an offset into its allocated portion of address space.

Required methods

fn debug_label(&self) -> String

Returns a label suitable for debug output.

Provided methods

fn read(&mut self, offset: BusAccessInfo, data: &mut [u8])

Reads at offset from this device

fn write(&mut self, offset: BusAccessInfo, data: &[u8])

Writes at offset into this device

fn config_register_write(
&mut self,
reg_idx: usize,
offset: u64,
data: &[u8]
) -> ConfigWriteResult

Sets a register in the configuration space. Only used by PCI.

reg_idx - The index of the config register to modify.

offset - Offset in to the register.

fn config_register_read(&self, reg_idx: usize) -> u32

Gets a register from the configuration space. Only used by PCI.

reg_idx - The index of the config register to read.

fn on_sandboxed(&mut self)

Invoked when the device is sandboxed.