Trait devices::PciDevice

pub trait PciDevice: Send {

    fn debug_label(&self) -> String;

    fn allocate_address(
        &mut self, 
        resources: &mut SystemAllocator
    ) -> Result<PciAddress, Error>;

    fn keep_rds(&self) -> Vec<RawDescriptor>;

    fn get_bar_configuration(
        &self, 
        bar_num: usize
    ) -> Option<PciBarConfiguration>;

    fn read_config_register(&self, reg_idx: usize) -> u32;

    fn write_config_register(
        &mut self, 
        reg_idx: usize, 
        offset: u64, 
        data: &[u8]
    );

    fn read_bar(&mut self, addr: u64, data: &mut [u8]);

    fn write_bar(&mut self, addr: u64, data: &[u8]);

    fn preferred_address(&self) -> Option<PciAddress> { ... }

    fn preferred_irq(&self) -> Option<(PciInterruptPin, u32)> { ... }

    fn assign_irq(
        &mut self, 
        _irq_evt: IrqLevelEvent, 
        _pin: PciInterruptPin, 
        _irq_num: u32
    ) { ... }

    fn allocate_io_bars(
        &mut self, 
        _resources: &mut SystemAllocator
    ) -> Result<Vec<BarRange>, Error> { ... }

    fn allocate_device_bars(
        &mut self, 
        _resources: &mut SystemAllocator
    ) -> Result<Vec<BarRange>, Error> { ... }

    fn register_device_capabilities(&mut self) -> Result<(), Error> { ... }

    fn ioevents(&self) -> Vec<(&Event, u64, Datamatch)> { ... }

    fn read_virtual_config_register(&self, _reg_idx: usize) -> u32 { ... }

    fn write_virtual_config_register(&mut self, _reg_idx: usize, _value: u32) { ... }

    fn on_device_sandboxed(&mut self) { ... }

    fn generate_acpi(&mut self, sdts: Vec<SDT>) -> Option<Vec<SDT>> { ... }

    fn destroy_device(&mut self) { ... }

    fn get_removed_children_devices(&self) -> Vec<PciAddress> { ... }

    fn get_new_pci_bus(&self) -> Option<Arc<Mutex<PciBus>>> { ... }

    fn configure_bridge_window(
        &mut self, 
        _resources: &mut SystemAllocator, 
        _bar_ranges: &[BarRange]
    ) -> Result<Vec<BarRange>, Error> { ... }

    fn set_subordinate_bus(&mut self, _bus_no: u8) { ... }

    fn supports_iommu(&self) -> bool { ... }

    fn set_iommu(&mut self, _iommu: IpcMemoryMapper) -> Result<()> { ... }

}

Required methods

fn debug_label(&self) -> String

Returns a label suitable for debug output.

fn allocate_address(
&mut self,
resources: &mut SystemAllocator
) -> Result<PciAddress, Error>

Allocate and return an unique bus, device and function number for this device. May be called multiple times; on subsequent calls, the device should return the same address it returned from the first call.

fn keep_rds(&self) -> Vec<RawDescriptor>

A vector of device-specific file descriptors that must be kept open after jailing. Must be called before the process is jailed.

fn get_bar_configuration(&self, bar_num: usize) -> Option<PciBarConfiguration>

Returns the configuration of a base address register, if present.

fn read_config_register(&self, reg_idx: usize) -> u32

Reads from a PCI configuration register.

reg_idx - PCI register index (in units of 4 bytes).

fn write_config_register(&mut self, reg_idx: usize, offset: u64, data: &[u8])

Writes to a PCI configuration register.

reg_idx - PCI register index (in units of 4 bytes).

offset - byte offset within 4-byte register.

data - The data to write.

fn read_bar(&mut self, addr: u64, data: &mut [u8])

Reads from a BAR region mapped in to the device.

addr - The guest address inside the BAR.

data - Filled with the data from addr.

fn write_bar(&mut self, addr: u64, data: &[u8])

Writes to a BAR region mapped in to the device.

addr - The guest address inside the BAR.

data - The data to write.

Provided methods

fn preferred_address(&self) -> Option<PciAddress>

Preferred PCI address for this device, if any.

fn preferred_irq(&self) -> Option<(PciInterruptPin, u32)>

Preferred IRQ for this device. The device may request a specific pin and IRQ number by returning a non-None value. Otherwise, an appropriate IRQ will be allocated automatically. The device’s assign_irq function will be called with its assigned IRQ either way.

fn assign_irq(
&mut self,
_irq_evt: IrqLevelEvent,
_pin: PciInterruptPin,
_irq_num: u32
)

Assign a legacy PCI IRQ to this device. The device may write to irq_evt to trigger an interrupt. When irq_resample_evt is signaled, the device should re-assert irq_evt if necessary.

fn allocate_io_bars(
&mut self,
_resources: &mut SystemAllocator
) -> Result<Vec<BarRange>, Error>

Allocates the needed IO BAR space using the allocate function which takes a size and returns an address. Returns a Vec of BarRange{addr, size, prefetchable}.

fn allocate_device_bars(
&mut self,
_resources: &mut SystemAllocator
) -> Result<Vec<BarRange>, Error>

Allocates the needed device BAR space. Returns a Vec of BarRange{addr, size, prefetchable}. Unlike MMIO BARs (see allocate_io_bars), device BARs are not expected to incur VM exits

these BARs represent normal memory.

fn register_device_capabilities(&mut self) -> Result<(), Error>

Register any capabilties specified by the device.

fn ioevents(&self) -> Vec<(&Event, u64, Datamatch)>

Gets a list of ioevents that should be registered with the running VM. The list is returned as a Vec of (event, addr, datamatch) tuples.

fn read_virtual_config_register(&self, _reg_idx: usize) -> u32

Reads from a virtual config register.

reg_idx - virtual config register index (in units of 4 bytes).

fn write_virtual_config_register(&mut self, _reg_idx: usize, _value: u32)

Writes to a virtual config register.

reg_idx - virtual config register index (in units of 4 bytes).

value - the value to be written.

fn on_device_sandboxed(&mut self)

Invoked when the device is sandboxed.

fn generate_acpi(&mut self, sdts: Vec<SDT>) -> Option<Vec<SDT>>

fn destroy_device(&mut self)

Invoked when the device is destroyed

fn get_removed_children_devices(&self) -> Vec<PciAddress>

Get the removed children devices under pci bridge

fn get_new_pci_bus(&self) -> Option<Arc<Mutex<PciBus>>>

Get the pci bus generated by this pci device

fn configure_bridge_window(
&mut self,
_resources: &mut SystemAllocator,
_bar_ranges: &[BarRange]
) -> Result<Vec<BarRange>, Error>

if device is a pci brdige, configure pci bridge window

fn set_subordinate_bus(&mut self, _bus_no: u8)

if device is a pci bridge, configure subordinate bus number

fn supports_iommu(&self) -> bool

Indicates whether the device supports IOMMU

fn set_iommu(&mut self, _iommu: IpcMemoryMapper) -> Result<()>

Sets the IOMMU for the device if supports_iommu()