Trait devices::PciDevice

pub trait PciDevice: Send {

    fn debug_label(&self) -> String;

    fn allocate_address(
        &mut self, 
        resources: &mut SystemAllocator
    ) -> Result<PciAddress, Error>;

    fn keep_rds(&self) -> Vec<RawDescriptor>;

    fn get_bar_configuration(
        &self, 
        bar_num: usize
    ) -> Option<PciBarConfiguration>;

    fn read_config_register(&self, reg_idx: usize) -> u32;

    fn write_config_register(
        &mut self, 
        reg_idx: usize, 
        offset: u64, 
        data: &[u8]
    );

    fn read_bar(&mut self, addr: u64, data: &mut [u8]);

    fn write_bar(&mut self, addr: u64, data: &[u8]);

    fn assign_irq(
        &mut self, 
        _irq_evt: &Event, 
        _irq_resample_evt: &Event, 
        _irq_num: Option<u32>
    ) -> Option<(u32, PciInterruptPin)> { ... }

    fn allocate_io_bars(
        &mut self, 
        _resources: &mut SystemAllocator
    ) -> Result<Vec<(u64, u64)>, Error> { ... }

    fn allocate_device_bars(
        &mut self, 
        _resources: &mut SystemAllocator
    ) -> Result<Vec<(u64, u64)>, Error> { ... }

    fn register_device_capabilities(&mut self) -> Result<(), Error> { ... }

    fn ioevents(&self) -> Vec<(&Event, u64, Datamatch)> { ... }

    fn on_device_sandboxed(&mut self) { ... }

    fn generate_acpi(&mut self, sdts: Vec<SDT>) -> Option<Vec<SDT>> { ... }

    fn destroy_device(&mut self) { ... }

    fn get_removed_children_devices(&self) -> Vec<PciAddress> { ... }

}

Required methods

fn debug_label(&self) -> String

Returns a label suitable for debug output.

fn allocate_address(
&mut self,
resources: &mut SystemAllocator
) -> Result<PciAddress, Error>

Allocate and return an unique bus, device and function number for this device.

fn keep_rds(&self) -> Vec<RawDescriptor>

A vector of device-specific file descriptors that must be kept open after jailing. Must be called before the process is jailed.

fn get_bar_configuration(&self, bar_num: usize) -> Option<PciBarConfiguration>

Returns the configuration of a base address register, if present.

fn read_config_register(&self, reg_idx: usize) -> u32

Reads from a PCI configuration register.

reg_idx - PCI register index (in units of 4 bytes).

fn write_config_register(&mut self, reg_idx: usize, offset: u64, data: &[u8])

Writes to a PCI configuration register.

reg_idx - PCI register index (in units of 4 bytes).

offset - byte offset within 4-byte register.

data - The data to write.

fn read_bar(&mut self, addr: u64, data: &mut [u8])

Reads from a BAR region mapped in to the device.

addr - The guest address inside the BAR.

data - Filled with the data from addr.

fn write_bar(&mut self, addr: u64, data: &[u8])

Writes to a BAR region mapped in to the device.

addr - The guest address inside the BAR.

data - The data to write.

Provided methods

fn assign_irq(
&mut self,
_irq_evt: &Event,
_irq_resample_evt: &Event,
_irq_num: Option<u32>
) -> Option<(u32, PciInterruptPin)>

Assign a legacy PCI IRQ to this device. The device may write to irq_evt to trigger an interrupt. When irq_resample_evt is signaled, the device should re-assert irq_evt if necessary. Optional irq_num can be used for default INTx allocation, device can overwrite it. If legacy INTx is used, function shall return requested IRQ number and PCI INTx pin.

fn allocate_io_bars(
&mut self,
_resources: &mut SystemAllocator
) -> Result<Vec<(u64, u64)>, Error>

Allocates the needed IO BAR space using the allocate function which takes a size and returns an address. Returns a Vec of (address, length) tuples.

fn allocate_device_bars(
&mut self,
_resources: &mut SystemAllocator
) -> Result<Vec<(u64, u64)>, Error>

Allocates the needed device BAR space. Returns a Vec of (address, length) tuples. Unlike MMIO BARs (see allocate_io_bars), device BARs are not expected to incur VM exits

these BARs represent normal memory.

fn register_device_capabilities(&mut self) -> Result<(), Error>

Register any capabilties specified by the device.

fn ioevents(&self) -> Vec<(&Event, u64, Datamatch)>

Gets a list of ioevents that should be registered with the running VM. The list is returned as a Vec of (event, addr, datamatch) tuples.

fn on_device_sandboxed(&mut self)

Invoked when the device is sandboxed.

fn generate_acpi(&mut self, sdts: Vec<SDT>) -> Option<Vec<SDT>>

fn destroy_device(&mut self)

Invoked when the device is destroyed

fn get_removed_children_devices(&self) -> Vec<PciAddress>

Get the removed children devices under pci bridge