Struct devices::vfio::VfioDevice

pub struct VfioDevice { /* fields omitted */ }

Vfio device for exposing regions which could be read/write to kernel vfio device.

Implementations

impl VfioDevice

pub fn new_passthrough<P: AsRef<Path>>(
sysfspath: &P,
vm: &impl Vm,
container: Arc<Mutex<VfioContainer>>,
iommu_enabled: bool
) -> Result<Self, VfioError>

Create a new vfio device, then guest read/write on this device could be transfered into kernel vfio. sysfspath specify the vfio device path in sys file system.

pub fn new<P: AsRef<Path>>(
sysfspath: &P,
container: Arc<Mutex<VfioContainer>>
) -> Result<Self, VfioError>

pub fn device_name(&self) -> &String

Returns PCI device name, formatted as BUS:DEVICE.FUNCTION string.

pub fn irq_enable(
&self,
descriptors: &[&Event],
index: u32
) -> Result<(), VfioError>

Enable vfio device’s irq and associate Irqfd Event with device. When MSIx is enabled, multi vectors will be supported, so descriptors is vector and the vector length is the num of MSIx vectors

pub fn resample_virq_enable(
&self,
descriptor: &Event,
index: u32
) -> Result<(), VfioError>

When intx is enabled, irqfd is used to trigger a level interrupt into guest, resample irqfd is used to get guest EOI notification. When host hw generates interrupt, vfio irq handler in host kernel receive and handle it, this handler disable hw irq first, then trigger irqfd to inject interrupt into guest. When resample irqfd is triggered by guest EOI, vfio kernel could enable hw irq, so hw could generate another interrupts. This function enable resample irqfd and let vfio kernel could get EOI notification.

descriptor: should be resample IrqFd.

pub fn irq_disable(&self, index: u32) -> Result<(), VfioError>

disable vfio device’s irq and disconnect Irqfd Event with device

pub fn irq_unmask(&self, index: u32) -> Result<(), VfioError>

Unmask vfio device irq

pub fn irq_mask(&self, index: u32) -> Result<(), VfioError>

Mask vfio device irq

pub fn get_irqs(&self) -> Result<Vec<VfioIrq>, VfioError>

Query interrupt information return: Vector of interrupts information, each of which contains flags and index

pub fn get_region_flags(&self, index: u32) -> u32

get a region’s flag the return’s value may conatin: VFIO_REGION_INFO_FLAG_READ: region supports read VFIO_REGION_INFO_FLAG_WRITE: region supports write VFIO_REGION_INFO_FLAG_MMAP: region supports mmap VFIO_REGION_INFO_FLAG_CAPS: region’s info supports caps

pub fn get_region_offset(&self, index: u32) -> u64

get a region’s offset return: Region offset from the start of vfio device descriptor

pub fn get_region_size(&self, index: u32) -> u64

get a region’s size return: Region size from the start of vfio device descriptor

pub fn get_region_count(&self) -> u32

get a number of regions return: Number of regions of vfio device descriptor

pub fn get_region_mmap(&self, index: u32) -> Vec<vfio_region_sparse_mmap_area>

get a region’s mmap info vector

pub fn get_cap_type_info(&self, type_: u32, sub_type: u32) -> Option<(u32, u64)>

find the specified cap type in device regions Input: type_: cap type sub_type: cap sub_type Output: None: device doesn’t have the specified cap type Some((bar_index, region_size)): device has the specified cap type, return region’s index and size

pub fn get_offset_for_addr(
&self,
addr: &VfioRegionAddr
) -> Result<u64, VfioError>

Returns file offset corresponding to the given VfioRegionAddr. The offset can be used when reading/writing the VFIO device’s FD directly.

pub fn region_read(&self, index: u32, buf: &mut [u8], addr: u64)

Read region’s data from VFIO device into buf index: region num buf: data destination and buf length is read size addr: offset in the region

pub fn region_read_from_addr<T: DataInit>(
&self,
addr: &VfioRegionAddr,
offset: u64
) -> T

Reads a value from the specified VfioRegionAddr.addr + offset.

pub fn region_write(&self, index: u32, buf: &[u8], addr: u64)

write the data from buf into a vfio device region index: region num buf: data src and buf length is write size addr: offset in the region

pub fn region_write_to_addr<T: DataInit>(
&self,
val: &T,
addr: &VfioRegionAddr,
offset: u64
)

Writes data into the specified VfioRegionAddr.addr + offset.

pub fn keep_rds(&self) -> Vec<RawDescriptor>

get vfio device’s descriptors which are passed into minijail process

pub unsafe fn vfio_dma_map(
&self,
iova: u64,
size: u64,
user_addr: u64,
write_en: bool
) -> Result<(), VfioError>

Add (iova, user_addr) map into vfio container iommu table

pub fn vfio_dma_unmap(&self, iova: u64, size: u64) -> Result<(), VfioError>

Remove (iova, user_addr) map from vfio container iommu table

pub fn device_file(&self) -> &File

Gets the vfio device backing File.

pub fn close(&self)

close vfio device