Function devices::virtio::console::handle_input

pub fn handle_input<I: SignalableInterrupt>(
    mem: &GuestMemory, 
    interrupt: &I, 
    buffer: &mut VecDeque<u8>, 
    receive_queue: &mut Queue
) -> Result<(), ConsoleError>

Checks for input from buffer and transfers it to the receive queue, if any.

Arguments

mem - The GuestMemory to write the data into

interrupt - SignalableInterrupt used to signal that the queue has been used

buffer - Ring buffer providing data to put into the guest

receive_queue - The receive virtio Queue