Module devices::virtio::fs::passthrough

Structs

Config

Options that configure the behavior of the file system.

PassthroughFs

A file system that simply “passes through” all requests it receives to the underlying file system. To keep the implementation simple it servers the contents of its root directory. Users that wish to serve only a specific directory should set up the environment so that that directory ends up as the root of the file system process. One way to accomplish this is via a combination of mount namespaces and the pivot_root system call.

Enums

CachePolicy

The caching policy that the file system should report to the FUSE client. By default the FUSE protocol uses close-to-open consistency. This means that any cached contents of the file are invalidated the next time that file is opened.

Functions

FS_IOC32_GETFLAGS

FS_IOC32_SETFLAGS

FS_IOC64_GETFLAGS

FS_IOC64_SETFLAGS

FS_IOC_ENABLE_VERITY

FS_IOC_FSGETXATTR

FS_IOC_FSSETXATTR

FS_IOC_GETFLAGS

FS_IOC_GET_ENCRYPTION_POLICY_EX

FS_IOC_MEASURE_VERITY

FS_IOC_SETFLAGS