Struct devices::virtio::fs::passthrough::PassthroughFs

pub struct PassthroughFs {
    inodes: Mutex<MultikeyBTreeMap<u64, InodeAltKey, Arc<InodeData>>>,
    next_inode: AtomicU64,
    handles: Mutex<BTreeMap<u64, Arc<HandleData>>>,
    next_handle: AtomicU64,
    proc: File,
    writeback: AtomicBool,
    zero_message_open: AtomicBool,
    zero_message_opendir: AtomicBool,
    cfg: Config,
}

A file system that simply “passes through” all requests it receives to the underlying file system. To keep the implementation simple it servers the contents of its root directory. Users that wish to serve only a specific directory should set up the environment so that that directory ends up as the root of the file system process. One way to accomplish this is via a combination of mount namespaces and the pivot_root system call.

Fields
inodes: Mutex<MultikeyBTreeMap<u64, InodeAltKey, Arc<InodeData>>>next_inode: AtomicU64handles: Mutex<BTreeMap<u64, Arc<HandleData>>>next_handle: AtomicU64proc: Filewriteback: AtomicBoolzero_message_open: AtomicBoolzero_message_opendir: AtomicBoolcfg: Config
Implementations

impl PassthroughFs

pub fn new(cfg: Config) -> Result<PassthroughFs>

pub fn cfg(&self) -> &Config

pub fn keep_rds(&self) -> Vec<RawDescriptor>

fn rewrite_xattr_name<'xattr>(&self, name: &'xattr CStr) -> Cow<'xattr, CStr>

fn find_inode(&self, inode: u64) -> Result<Arc<InodeData>>

fn find_handle(&self, handle: u64, inode: u64) -> Result<Arc<HandleData>>

fn open_fd(&self, fd: RawDescriptor, flags: i32) -> Result<File>

fn open_inode(&self, inode: &InodeData, flags: i32) -> Result<File>

fn add_entry(&self, f: File, st: stat64, open_flags: c_int) -> Entry

fn ascii_casefold_lookup(
&self,
parent: &InodeData,
name: &[u8]
) -> Result<Entry>

fn do_lookup(&self, parent: &InodeData, name: &CStr) -> Result<Entry>

fn do_open(&self, inode: u64, flags: u32) -> Result<(Option<u64>, OpenOptions)>

fn do_release(&self, inode: u64, handle: u64) -> Result<()>

fn do_getattr(&self, inode: &InodeData) -> Result<(stat64, Duration)>

fn do_unlink(&self, parent: &InodeData, name: &CStr, flags: c_int) -> Result<()>

fn do_fsync<F: AsRawDescriptor>(&self, file: &F, datasync: bool) -> Result<()>

fn with_proc_chdir<F, T>(&self, f: F) -> T where
F: FnOnce() -> T,

fn do_getxattr(
&self,
inode: &InodeData,
name: &CStr,
value: &mut [u8]
) -> Result<usize>

fn get_encryption_policy_ex<R: Read>(
&self,
inode: u64,
handle: u64,
r: R
) -> Result<IoctlReply>

fn get_fsxattr(&self, inode: u64, handle: u64) -> Result<IoctlReply>

fn set_fsxattr<R: Read>(
&self,
ctx: Context,
inode: u64,
handle: u64,
r: R
) -> Result<IoctlReply>

fn get_flags(&self, inode: u64, handle: u64) -> Result<IoctlReply>

fn set_flags<R: Read>(
&self,
ctx: Context,
inode: u64,
handle: u64,
r: R
) -> Result<IoctlReply>

fn enable_verity<R: Read>(
&self,
inode: u64,
handle: u64,
r: R
) -> Result<IoctlReply>

fn measure_verity<R: Read>(
&self,
inode: u64,
handle: u64,
r: R,
out_size: u32
) -> Result<IoctlReply>