Struct devices::virtio::fs::passthrough::PassthroughFs

pub struct PassthroughFs { /* fields omitted */ }

A file system that simply “passes through” all requests it receives to the underlying file system. To keep the implementation simple it servers the contents of its root directory. Users that wish to serve only a specific directory should set up the environment so that that directory ends up as the root of the file system process. One way to accomplish this is via a combination of mount namespaces and the pivot_root system call.

Implementations

impl PassthroughFs

pub fn new(cfg: Config) -> Result<PassthroughFs>

pub fn cfg(&self) -> &Config

pub fn keep_rds(&self) -> Vec<RawDescriptor>