Module devices::virtio::gpu

Re-exports

pub use self::protocol::virtio_gpu_config;

pub use self::protocol::VIRTIO_GPU_F_CONTEXT_INIT;

pub use self::protocol::VIRTIO_GPU_F_CREATE_GUEST_HANDLE;

pub use self::protocol::VIRTIO_GPU_F_EDID;

pub use self::protocol::VIRTIO_GPU_F_RESOURCE_BLOB;

pub use self::protocol::VIRTIO_GPU_F_RESOURCE_SYNC;

pub use self::protocol::VIRTIO_GPU_F_RESOURCE_UUID;

pub use self::protocol::VIRTIO_GPU_F_VIRGL;

pub use super::device_constants::gpu::QUEUE_SIZES;

Modules

edid

Implementation of the EDID specification provided by software. EDID spec: https://glenwing.github.io/docs/VESA-EEDID-A2.pdf

parameters

Definitions and utilities for GPU related parameters.

protocol

sys

virtio_gpu

Structs

EventDevice

Encapsulates a virtual event device, such as a mouse or keyboard

FenceDescriptor

FenceState

Frontend

Gpu

GpuDisplayParameters

GpuParameters

LocalQueueReader

ReturnDescriptor

SharedQueueReader

VirtioScanoutBlobData

Worker

Enums

DisplayBackend

Indicates a backend that should be tried for the gpu to use for display.

GpuDisplayMode

GpuMode

VirtioGpuRing

Constants

DEFAULT_DISPLAY_HEIGHT

DEFAULT_DISPLAY_WIDTH

DEFAULT_REFRESH_RATE

FENCE_POLL_INTERVAL

VIRTIO_GPU_SHM_ID_HOST_VISIBLE

Traits

QueueReader

Functions

build

Initializes the virtio_gpu state tracker.

create_fence_handler

Create a handler that writes into the completed fence queue