Module devices::virtio::gpu

Modules

parameters

Definitions and utilities for GPU related parameters.

protocol

udmabuf

udmabuf_bindings

virtio_gpu

Structs

EventDevice

Encapsulates a virtual event device, such as a mouse or keyboard

FenceDescriptor

FenceState

Frontend

Gpu

GpuDisplayParameters

GpuParameters

LocalQueueReader

ReturnDescriptor

SharedQueueReader

VirtioScanoutBlobData

Worker

virtio_gpu_config

Enums

DisplayBackend

Indicates a backend that should be tried for the gpu to use for display.

GpuMode

VirtioGpuRing

Constants

DEFAULT_DISPLAY_HEIGHT

DEFAULT_DISPLAY_WIDTH

FENCE_POLL_INTERVAL

GPU_BAR_NUM

GPU_BAR_OFFSET

QUEUE_SIZES

VIRTIO_GPU_F_CONTEXT_INIT

VIRTIO_GPU_F_CREATE_GUEST_HANDLE

VIRTIO_GPU_F_EDID

VIRTIO_GPU_F_RESOURCE_BLOB

VIRTIO_GPU_F_RESOURCE_SYNC

VIRTIO_GPU_F_RESOURCE_UUID

VIRTIO_GPU_F_VIRGL

VIRTIO_GPU_SHM_ID_HOST_VISIBLE

Traits

QueueReader

Functions

build

Initializes the virtio_gpu state tracker.

create_fence_handler

Create a handler that writes into the completed fence queue