Module devices::virtio::gpu

Structs

EventDevice

Encapsulates a virtual event device, such as a mouse or keyboard

FenceState

Frontend

Gpu

GpuDisplayParameters

GpuParameters

GpuRenderServerParameters

ReturnDescriptor

VirtioScanoutBlobData

virtio_gpu_config

Enums

DisplayBackend

Indicates a backend that should be tried for the gpu to use for display.

GpuMode

Constants

DEFAULT_DISPLAY_HEIGHT

DEFAULT_DISPLAY_WIDTH

FENCE_POLL_INTERVAL

GPU_BAR_NUM

GPU_BAR_OFFSET

GPU_BAR_SIZE

QUEUE_SIZES

VIRTIO_GPU_F_CONTEXT_INIT

VIRTIO_GPU_F_CREATE_GUEST_HANDLE

VIRTIO_GPU_F_EDID

VIRTIO_GPU_F_RESOURCE_BLOB

VIRTIO_GPU_F_RESOURCE_SYNC

VIRTIO_GPU_F_RESOURCE_UUID

VIRTIO_GPU_F_VIRGL

VIRTIO_GPU_SHM_ID_HOST_VISIBLE

Traits

QueueReader

Functions

create_fence_handler

Create a handler that writes into the completed fence queue