Struct devices::virtio::gpu::virtio_gpu::VirtioGpu

pub struct VirtioGpu {
    display: Rc<RefCell<GpuDisplay>>,
    scanouts: Vec<VirtioGpuScanout>,
    cursor_scanout: VirtioGpuScanout,
    event_devices: Map<u32, u32>,
    mapper: Box<dyn SharedMemoryMapper>,
    map_request: Arc<Mutex<Option<ExternalMapping>>>,
    rutabaga: Rutabaga,
    resources: Map<u32, VirtioGpuResource>,
    external_blob: bool,
    refresh_rate: u32,
    udmabuf_driver: Option<UdmabufDriver>,
}

Handles functionality related to displays, input events and hypervisor memory management.

Fields
display: Rc<RefCell<GpuDisplay>>scanouts: Vec<VirtioGpuScanout>cursor_scanout: VirtioGpuScanoutevent_devices: Map<u32, u32>mapper: Box<dyn SharedMemoryMapper>map_request: Arc<Mutex<Option<ExternalMapping>>>rutabaga: Rutabagaresources: Map<u32, VirtioGpuResource>external_blob: boolrefresh_rate: u32udmabuf_driver: Option<UdmabufDriver>
Implementations

impl VirtioGpu

pub fn new(
display: GpuDisplay,
display_params: Vec<GpuDisplayParameters>,
rutabaga_builder: RutabagaBuilder,
event_devices: Vec<EventDevice>,
mapper: Box<dyn SharedMemoryMapper>,
map_request: Arc<Mutex<Option<ExternalMapping>>>,
external_blob: bool,
udmabuf: bool,
fence_handler: RutabagaFenceHandler,
render_server_fd: Option<SafeDescriptor>
) -> Option<VirtioGpu>

Creates a new instance of the VirtioGpu state tracker.

pub fn import_event_device(
&mut self,
event_device: EventDevice,
scanout_id: u32
) -> Result<GpuResponse, GpuResponse>

Imports the event device

pub fn display(&mut self) -> &Rc<RefCell<GpuDisplay>>

Gets a reference to the display passed into new.

pub fn display_info(&self) -> Vec<(u32, u32)>

Gets the list of supported display resolutions as a slice of (width, height) tuples.

pub fn process_display(&mut self) -> bool

Processes the internal display events and returns true if any display was closed.

pub fn set_scanout(
&mut self,
scanout_id: u32,
resource_id: u32,
scanout_data: Option<VirtioScanoutBlobData>
) -> Result<GpuResponse, GpuResponse>

Sets the given resource id as the source of scanout to the display.

pub fn flush_resource(
&mut self,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

If the resource is the scanout resource, flush it to the display.

pub fn update_cursor(
&mut self,
resource_id: u32,
scanout_id: u32,
x: u32,
y: u32
) -> Result<GpuResponse, GpuResponse>

Updates the cursor’s memory to the given resource_id, and sets its position to the given coordinates.

pub fn move_cursor(
&mut self,
_scanout_id: u32,
x: u32,
y: u32
) -> Result<GpuResponse, GpuResponse>

Moves the cursor’s position to the given coordinates.

pub fn resource_assign_uuid(
&self,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

Returns a uuid for the resource.

pub fn export_resource(&mut self, resource_id: u32) -> ResourceResponse

If supported, export the resource with the given resource_id to a file.

pub fn export_fence(&self, fence_id: u32) -> ResourceResponse

If supported, export the fence with the given fence_id to a file.

pub fn get_capset_info(&self, index: u32) -> Result<GpuResponse, GpuResponse>

Gets rutabaga’s capset information associated with index.

pub fn get_capset(
&self,
capset_id: u32,
version: u32
) -> Result<GpuResponse, GpuResponse>

Gets a capset from rutabaga.

pub fn force_ctx_0(&self)

Forces rutabaga to use it’s default context.

pub fn create_fence(
&mut self,
rutabaga_fence: RutabagaFence
) -> Result<GpuResponse, GpuResponse>

Creates a fence with the RutabagaFence that can be used to determine when the previous command completed.

pub fn needs_fence_poll(&mut self) -> bool

pub fn fence_poll(&mut self) -> Vec<RutabagaFence>

Returns an array of RutabagaFence, describing completed fences.

pub fn event_poll(&self)

Polls the Rutabaga backend.

pub fn poll_descriptor(&self) -> Option<SafeDescriptor>

Gets a pollable eventfd that signals the device to wakeup and poll the Rutabaga backend.

pub fn resource_create_3d(
&mut self,
resource_id: u32,
resource_create_3d: ResourceCreate3D
) -> Result<GpuResponse, GpuResponse>

Creates a 3D resource with the given properties and resource_id.

pub fn attach_backing(
&mut self,
resource_id: u32,
mem: &GuestMemory,
vecs: Vec<(GuestAddress, usize)>
) -> Result<GpuResponse, GpuResponse>

Attaches backing memory to the given resource, represented by a Vec of (address, size) tuples in the guest’s physical address space. Converts to RutabageIovec from the memory mapping.

pub fn detach_backing(
&mut self,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

Detaches any previously attached iovecs from the resource.

pub fn unref_resource(
&mut self,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

Releases guest kernel reference on the resource.

pub fn transfer_write(
&mut self,
ctx_id: u32,
resource_id: u32,
transfer: Transfer3D
) -> Result<GpuResponse, GpuResponse>

Copies data to host resource from the attached iovecs. Can also be used to flush caches.

pub fn transfer_read(
&mut self,
ctx_id: u32,
resource_id: u32,
transfer: Transfer3D,
buf: Option<VolatileSlice<'_>>
) -> Result<GpuResponse, GpuResponse>

Copies data from the host resource to:

To the optional volatile slice

To the host resource’s attached iovecs

Can also be used to invalidate caches.

pub fn resource_create_blob(
&mut self,
ctx_id: u32,
resource_id: u32,
resource_create_blob: ResourceCreateBlob,
vecs: Vec<(GuestAddress, usize)>,
mem: &GuestMemory
) -> Result<GpuResponse, GpuResponse>

Creates a blob resource using rutabaga.

pub fn resource_map_blob(
&mut self,
resource_id: u32,
offset: u64
) -> Result<GpuResponse, GpuResponse>

Uses the hypervisor to map the rutabaga blob resource.

pub fn resource_unmap_blob(
&mut self,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

Uses the hypervisor to unmap the blob resource.

pub fn get_edid(&self, _scanout_id: u32) -> Result<GpuResponse, GpuResponse>

Gets the EDID for the specified scanout ID. We return a virtual EDID that is identical for all scanouts.

pub fn create_context(
&mut self,
ctx_id: u32,
context_init: u32,
context_name: Option<&str>
) -> Result<GpuResponse, GpuResponse>

Creates a rutabaga context.

pub fn destroy_context(
&mut self,
ctx_id: u32
) -> Result<GpuResponse, GpuResponse>

Destroys a rutabaga context.

pub fn context_attach_resource(
&mut self,
ctx_id: u32,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

Attaches a resource to a rutabaga context.

pub fn context_detach_resource(
&mut self,
ctx_id: u32,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>

Detaches a resource from a rutabaga context.

pub fn submit_command(
&mut self,
ctx_id: u32,
commands: &mut [u8]
) -> Result<GpuResponse, GpuResponse>

Submits a command buffer to a rutabaga context.

fn result_from_query(&mut self, resource_id: u32) -> GpuResponse

fn update_scanout_resource(
&mut self,
scanout_type: SurfaceType,
scanout_id: u32,
scanout_data: Option<VirtioScanoutBlobData>,
resource_id: u32
) -> Result<GpuResponse, GpuResponse>