Module devices::virtio::input::event_source

Structs

EvdevEventSource

Encapsulates an event device node as an event source

EventSourceImpl

Encapsulates implementation details common to all kinds of event sources.

SocketEventSource

Encapsulates a (unix) socket as an event source.

Enums

EventType

Traits

EventSource

Encapsulates a socket or device node into an abstract event source, providing a common interface. It supports read and write operations to provide and accept events just like an event device node would, except that it handles virtio_input_event instead of input_event structures. It’s necessary to call receive_events() before events are available for read.