Module devices::virtio::iommu

Modules

ipc_memory_mapper

Provide utility to communicate with an iommu in another process

memory_mapper

MemoryMapper trait and basic impl for virtio-iommu implementation

memory_util

Utilities to access GuestMemory with IO virtual addresses and iommu

protocol

This file was generated by the following commands and modified manually.

sys

Structs

Iommu

Virtio device for IOMMU memory management.

State

VirtioIommuViotHeader

VirtioIommuViotPciRangeNode

VirtioIommuViotVirtioPciNode

Enums

IommuError

Constants

IOMMU_PROBE_SIZE

NUM_QUEUES

QUEUE_SIZE

QUEUE_SIZES

VIRTIO_IOMMU_VIOT_NODE_PCI_RANGE

VIRTIO_IOMMU_VIOT_NODE_VIRTIO_IOMMU_PCI

Functions

request_queue

run

Type Definitions

Result