Struct devices::virtio::Reader

pub struct Reader { /* fields omitted */ }

Provides high-level interface over the sequence of memory regions defined by readable descriptors in the descriptor chain.

Note that virtio spec requires driver to place any device-writable descriptors after any device-readable descriptors (2.6.4.2 in Virtio Spec v1.1). Reader will skip iterating over descriptor chain when first writable descriptor is encountered.

Implementations

impl Reader

pub fn new(
mem: GuestMemory,
desc_chain: DescriptorChain
) -> Result<Reader, Error>

Construct a new Reader wrapper over desc_chain.

pub fn read_obj<T: DataInit>(&mut self) -> Result<T>

Reads an object from the descriptor chain buffer.

pub fn collect<C: FromIterator<Result<T>>, T: DataInit>(&mut self) -> C

Reads objects by consuming all the remaining data in the descriptor chain buffer and returns them as a collection. Returns an error if the size of the remaining data is indivisible by the size of an object of type T.

pub fn iter<T: DataInit>(&mut self) -> ReaderIterator<'_, T>

Creates an iterator for sequentially reading DataInit objects from the Reader. Unlike collect, this doesn’t consume all the remaining data in the Reader and doesn’t require the objects to be stored in a separate collection.

pub fn read_to_volatile_slice(&mut self, slice: VolatileSlice<'_>) -> usize

Reads data into a volatile slice up to the minimum of the slice’s length or the number of bytes remaining. Returns the number of bytes read.

pub fn read_to<F: FileReadWriteVolatile>(
&mut self,
dst: F,
count: usize
) -> Result<usize>

Reads data from the descriptor chain buffer into a file descriptor. Returns the number of bytes read from the descriptor chain buffer. The number of bytes read can be less than count if there isn’t enough data in the descriptor chain buffer.

pub fn read_to_at<F: FileReadWriteAtVolatile>(
&mut self,
dst: F,
count: usize,
off: u64
) -> Result<usize>

Reads data from the descriptor chain buffer into a File at offset off. Returns the number of bytes read from the descriptor chain buffer. The number of bytes read can be less than count if there isn’t enough data in the descriptor chain buffer.

pub fn read_exact_to<F: FileReadWriteVolatile>(
&mut self,
dst: F,
count: usize
) -> Result<()>

Reads data from the descriptor chain similar to ‘read_to’ except reading ‘count’ or returning an error if ‘count’ bytes can’t be read.

pub fn read_exact_to_at<F: FileReadWriteAtVolatile>(
&mut self,
dst: F,
count: usize,
off: u64
) -> Result<()>

Reads data from the descriptor chain similar to ‘read_to_at’ except reading ‘count’ or returning an error if ‘count’ bytes can’t be read.

pub async fn read_to_at_fut<F: AsyncDisk + ?Sized>(
&mut self,
dst: &F,
count: usize,
off: u64
) -> Result<usize>

Reads data from the descriptor chain buffer into an AsyncDisk at offset off. Returns the number of bytes read from the descriptor chain buffer. The number of bytes read can be less than count if there isn’t enough data in the descriptor chain buffer.

pub async fn read_exact_to_at_fut<F: AsyncDisk + ?Sized>(
&mut self,
dst: &F,
count: usize,
off: u64
) -> Result<()>

Reads exactly count bytes from the chain to the disk asynchronously or returns an error if not enough data can be read.

pub fn available_bytes(&self) -> usize

Returns number of bytes available for reading. May return an error if the combined lengths of all the buffers in the DescriptorChain would cause an integer overflow.

pub fn bytes_read(&self) -> usize

Returns number of bytes already read from the descriptor chain buffer.

pub fn get_remaining(&self) -> SmallVec<[VolatileSlice<'_>; 16]>

Returns a &[VolatileSlice] that represents all the remaining data in this Reader. Calling this method does not actually consume any data from the Reader and callers should call consume to advance the Reader.

pub fn consume(&mut self, amt: usize)

Consumes amt bytes from the underlying descriptor chain. If amt is larger than the remaining data left in this Reader, then all remaining data will be consumed.

pub fn split_at(&mut self, offset: usize) -> Reader

Splits this Reader into two at the given offset in the DescriptorChain buffer. After the split, self will be able to read up to offset bytes while the returned Reader can read up to available_bytes() - offset bytes. If offset > self.available_bytes(), then the returned Reader will not be able to read any bytes.