Struct devices::virtio::Writer

pub struct Writer { /* fields omitted */ }

Provides high-level interface over the sequence of memory regions defined by writable descriptors in the descriptor chain.

Note that virtio spec requires driver to place any device-writable descriptors after any device-readable descriptors (2.6.4.2 in Virtio Spec v1.1). Writer will start iterating the descriptors from the first writable one and will assume that all following descriptors are writable.

Implementations

impl Writer

pub fn new(
mem: GuestMemory,
desc_chain: DescriptorChain
) -> Result<Writer, Error>

Construct a new Writer wrapper over desc_chain.

pub fn write_obj<T: DataInit>(&mut self, val: T) -> Result<()>

Writes an object to the descriptor chain buffer.

pub fn write_iter<T: DataInit, I: Iterator<Item = T>>(
&mut self,
iter: I
) -> Result<()>

Writes all objects produced by iter into the descriptor chain buffer. Unlike consume, this doesn’t require the values to be stored in an intermediate collection first. It also allows callers to choose which elements in a collection to write, for example by using the filter or take methods of the Iterator trait.

pub fn consume<T: DataInit, C: IntoIterator<Item = T>>(
&mut self,
vals: C
) -> Result<()>

Writes a collection of objects into the descriptor chain buffer.

pub fn available_bytes(&self) -> usize

Returns number of bytes available for writing. May return an error if the combined lengths of all the buffers in the DescriptorChain would cause an overflow.

pub fn write_from_volatile_slice(&mut self, slice: VolatileSlice<'_>) -> usize

Reads data into a volatile slice up to the minimum of the slice’s length or the number of bytes remaining. Returns the number of bytes read.

pub fn write_from<F: FileReadWriteVolatile>(
&mut self,
src: F,
count: usize
) -> Result<usize>

Writes data to the descriptor chain buffer from a file descriptor. Returns the number of bytes written to the descriptor chain buffer. The number of bytes written can be less than count if there isn’t enough data in the descriptor chain buffer.

pub fn write_from_at<F: FileReadWriteAtVolatile>(
&mut self,
src: F,
count: usize,
off: u64
) -> Result<usize>

Writes data to the descriptor chain buffer from a File at offset off. Returns the number of bytes written to the descriptor chain buffer. The number of bytes written can be less than count if there isn’t enough data in the descriptor chain buffer.

pub fn write_all_from<F: FileReadWriteVolatile>(
&mut self,
src: F,
count: usize
) -> Result<()>

pub fn write_all_from_at<F: FileReadWriteAtVolatile>(
&mut self,
src: F,
count: usize,
off: u64
) -> Result<()>

pub async fn write_from_at_fut<F: AsyncDisk + ?Sized>(
&mut self,
src: &F,
count: usize,
off: u64
) -> Result<usize>

Writes data to the descriptor chain buffer from an AsyncDisk at offset off. Returns the number of bytes written to the descriptor chain buffer. The number of bytes written can be less than count if there isn’t enough data in the descriptor chain buffer.

pub async fn write_all_from_at_fut<F: AsyncDisk + ?Sized>(
&mut self,
src: &F,
count: usize,
off: u64
) -> Result<()>

pub fn bytes_written(&self) -> usize

Returns number of bytes already written to the descriptor chain buffer.

pub fn get_remaining(&self) -> SmallVec<[VolatileSlice<'_>; 16]>

Returns a &[VolatileSlice] that represents all the remaining data in this Writer. Calling this method does not actually advance the current position of the Writer in the buffer and callers should call consume_bytes to advance the Writer. Not calling consume_bytes with the amount of data copied into the returned VolatileSlices will result in that that data being overwritten the next time data is written into the Writer.

pub fn consume_bytes(&mut self, amt: usize)

Consumes amt bytes from the underlying descriptor chain. If amt is larger than the remaining data left in this Reader, then all remaining data will be consumed.

pub fn split_at(&mut self, offset: usize) -> Writer

Splits this Writer into two at the given offset in the DescriptorChain buffer. After the split, self will be able to write up to offset bytes while the returned Writer can write up to available_bytes() - offset bytes. If offset > self.available_bytes(), then the returned Writer will not be able to write any bytes.