Trait devices::virtio::VirtioDevice

pub trait VirtioDevice: Send {

    fn keep_rds(&self) -> Vec<RawDescriptor>;

    fn device_type(&self) -> u32;

    fn queue_max_sizes(&self) -> &[u16];

    fn activate(
        &mut self, 
        mem: GuestMemory, 
        interrupt: Interrupt, 
        queues: Vec<Queue>, 
        queue_evts: Vec<Event>
    );

    fn debug_label(&self) -> String { ... }

    fn features(&self) -> u64 { ... }

    fn ack_features(&mut self, value: u64) { ... }

    fn read_config(&self, offset: u64, data: &mut [u8]) { ... }

    fn write_config(&mut self, offset: u64, data: &[u8]) { ... }

    fn reset(&mut self) -> bool { ... }

    fn get_device_bars(
        &mut self, 
        _address: PciAddress
    ) -> Vec<PciBarConfiguration> { ... }

    fn get_device_caps(&self) -> Vec<Box<dyn PciCapability>> { ... }

    fn on_device_sandboxed(&mut self) { ... }

    fn control_notify(&self, _behavior: MsixStatus) { ... }

    fn generate_acpi(
        &mut self, 
        _pci_address: &Option<PciAddress>, 
        sdts: Vec<SDT>
    ) -> Option<Vec<SDT>> { ... }

    fn read_bar(&mut self, _bar_index: usize, _offset: u64, _data: &mut [u8]) { ... }

    fn write_bar(&mut self, _bar_index: usize, _offset: u64, _data: &[u8]) { ... }

}

Trait for virtio devices to be driven by a virtio transport.

The lifecycle of a virtio device is to be moved to a virtio transport, which will then query the device. Once the guest driver has configured the device, VirtioDevice::activate will be called and all the events, memory, and queues for device operation will be moved into the device. Optionally, a virtio device can implement device reset in which it returns said resources and resets its internal.

Required methods

fn keep_rds(&self) -> Vec<RawDescriptor>

A vector of device-specific file descriptors that must be kept open after jailing. Must be called before the process is jailed.

fn device_type(&self) -> u32

The virtio device type.

fn queue_max_sizes(&self) -> &[u16]

The maximum size of each queue that this device supports.

fn activate(
&mut self,
mem: GuestMemory,
interrupt: Interrupt,
queues: Vec<Queue>,
queue_evts: Vec<Event>
)

Activates this device for real usage.

Provided methods

fn debug_label(&self) -> String

Returns a label suitable for debug output.

fn features(&self) -> u64

The set of feature bits that this device supports in addition to the base features.

fn ack_features(&mut self, value: u64)

Acknowledges that this set of features should be enabled.

fn read_config(&self, offset: u64, data: &mut [u8])

Reads this device configuration space at offset.

fn write_config(&mut self, offset: u64, data: &[u8])

Writes to this device configuration space at offset.

fn reset(&mut self) -> bool

Optionally deactivates this device. If the reset method is not able to reset the virtio device, or the virtio device model doesn’t implement the reset method, a false value is returned to indicate the reset is not successful. Otherwise a true value should be returned.

fn get_device_bars(&mut self, _address: PciAddress) -> Vec<PciBarConfiguration>

Returns any additional BAR configuration required by the device.

fn get_device_caps(&self) -> Vec<Box<dyn PciCapability>>

Returns any additional capabiltiies required by the device.

fn on_device_sandboxed(&mut self)

Invoked when the device is sandboxed.

fn control_notify(&self, _behavior: MsixStatus)

fn generate_acpi(
&mut self,
_pci_address: &Option<PciAddress>,
sdts: Vec<SDT>
) -> Option<Vec<SDT>>

fn read_bar(&mut self, _bar_index: usize, _offset: u64, _data: &mut [u8])

Reads from a BAR region mapped in to the device.

addr - The guest address inside the BAR.

data - Filled with the data from addr.

fn write_bar(&mut self, _bar_index: usize, _offset: u64, _data: &[u8])

Writes to a BAR region mapped in to the device.

addr - The guest address inside the BAR.

data - The data to write.