Module devices::virtio::vhost::user::device

Functions

run_block_device

Starts a vhost-user block device. Returns an error if the given args is invalid or the device fails to run.

run_console_device

Starts a vhost-user console device. Returns an error if the given args is invalid or the device fails to run.

run_fs_device

Starts a vhost-user fs device. Returns an error if the given args is invalid or the device fails to run.

run_gpu_device

run_net_device

Starts a vhost-user net device. Returns an error if the given args is invalid or the device fails to run.

run_vsock_device

Returns an error if the given args is invalid or the device fails to run.

run_wl_device

Starts a vhost-user wayland device. Returns an error if the given args is invalid or the device fails to run.