Trait devices::virtio::vhost::user::device::handler::VhostUserBackend

pub trait VhostUserBackend {

    fn max_queue_num(&self) -> usize;

    fn max_vring_len(&self) -> u16;

    fn features(&self) -> u64;

    fn ack_features(&mut self, value: u64) -> Result<()>;

    fn acked_features(&self) -> u64;

    fn protocol_features(&self) -> VhostUserProtocolFeatures;

    fn ack_protocol_features(&mut self, _value: u64) -> Result<()>;

    fn acked_protocol_features(&self) -> u64;

    fn read_config(&self, offset: u64, dst: &mut [u8]);

    fn start_queue(
        &mut self, 
        idx: usize, 
        queue: Queue, 
        mem: GuestMemory, 
        doorbell: Arc<Mutex<Doorbell>>, 
        kick_evt: Event
    ) -> Result<()>;

    fn stop_queue(&mut self, idx: usize);

    fn reset(&mut self);

    fn write_config(&self, _offset: u64, _data: &[u8]) { ... }

    fn set_device_request_channel(&mut self, _channel: File) -> Result<()> { ... }

}

Trait for vhost-user backend.

Required methods

fn max_queue_num(&self) -> usize

The maximum number of queues that this backend can manage.

fn max_vring_len(&self) -> u16

The maximum length that virtqueues can take with this backend.

fn features(&self) -> u64

The set of feature bits that this backend supports.

fn ack_features(&mut self, value: u64) -> Result<()>

Acknowledges that this set of features should be enabled.

fn acked_features(&self) -> u64

Returns the set of enabled features.

fn protocol_features(&self) -> VhostUserProtocolFeatures

The set of protocol feature bits that this backend supports.

fn ack_protocol_features(&mut self, _value: u64) -> Result<()>

Acknowledges that this set of protocol features should be enabled.

fn acked_protocol_features(&self) -> u64

Returns the set of enabled protocol features.

fn read_config(&self, offset: u64, dst: &mut [u8])

Reads this device configuration space at offset.

fn start_queue(
&mut self,
idx: usize,
queue: Queue,
mem: GuestMemory,
doorbell: Arc<Mutex<Doorbell>>,
kick_evt: Event
) -> Result<()>

Indicates that the backend should start processing requests for virtio queue number idx. This method must not block the current thread so device backends should either spawn an async task or another thread to handle messages from the Queue.

fn stop_queue(&mut self, idx: usize)

Indicates that the backend should stop processing requests for virtio queue number idx.

fn reset(&mut self)

Resets the vhost-user backend.

Provided methods

fn write_config(&self, _offset: u64, _data: &[u8])

writes data to this device’s configuration space at offset.

fn set_device_request_channel(&mut self, _channel: File) -> Result<()>

Sets the channel for device-specific communication.