Enum devices::virtio::vhost::user::vmm::Error

pub enum Error {

    CopyConfig(Error),
    CreateEvent(Error),
    GetConfig(VhostError),
    GetFeatures(VhostError),
    GetHostAddress(GuestMemoryError),
    GetProtocolFeatures(VhostError),
    GetQueueNum(VhostError),
    GetVringBase(VhostError),
    InvalidConfigOffset {
        data_len: u64,
        offset: u64,
        config_len: u64,
    },
    MsixConfigUnavailable,
    MsixIrqfdUnavailable,
    ResetOwner(VhostError),
    SetConfig(VhostError),
    SetDeviceRequestChannel(VhostError),
    SetFeatures(VhostError),
    SetMemTable(VhostError),
    SetOwner(VhostError),
    SetProtocolFeatures(VhostError),
    SetVringAddr(VhostError),
    SetVringBase(VhostError),
    SetVringCall(VhostError),
    SetVringEnable(VhostError),
    SetVringKick(VhostError),
    SetVringNum(VhostError),
    SocketConnect(Error),
    SocketConnectOnMasterCreate(VhostError),
    TagTooLong {
        len: usize,
        max: usize,
    },

}

Variants

CopyConfig(Error)

Failed to copy config to a buffer.

Tuple Fields of CopyConfig

0: Error

CreateEvent(Error)

Failed to create base::Event.

Tuple Fields of CreateEvent

0: Error

GetConfig(VhostError)

Failed to get config.

Tuple Fields of GetConfig

0: VhostError

GetFeatures(VhostError)

Failed to get features.

Tuple Fields of GetFeatures

0: VhostError

GetHostAddress(GuestMemoryError)

Failed to get host address.

Tuple Fields of GetHostAddress

0: GuestMemoryError

GetProtocolFeatures(VhostError)

Failed to get protocol features.

Tuple Fields of GetProtocolFeatures

0: VhostError

GetQueueNum(VhostError)

Failed to get number of queues.

Tuple Fields of GetQueueNum

0: VhostError

GetVringBase(VhostError)

Failed to get vring base offset.

Tuple Fields of GetVringBase

0: VhostError

InvalidConfigOffset

Invalid config offset is given.

Fields of InvalidConfigOffset

data_len: u64offset: u64config_len: u64

MsixConfigUnavailable

MSI-X config is unavailable.

MsixIrqfdUnavailable

MSI-X irqfd is unavailable.

ResetOwner(VhostError)

Failed to reset owner.

Tuple Fields of ResetOwner

0: VhostError

SetConfig(VhostError)

Failed to set config.

Tuple Fields of SetConfig

0: VhostError

SetDeviceRequestChannel(VhostError)

Failed to set device request channel.

Tuple Fields of SetDeviceRequestChannel

0: VhostError

SetFeatures(VhostError)

Failed to set features.

Tuple Fields of SetFeatures

0: VhostError

SetMemTable(VhostError)

Failed to set memory map regions.

Tuple Fields of SetMemTable

0: VhostError

SetOwner(VhostError)

Failed to set owner.

Tuple Fields of SetOwner

0: VhostError

SetProtocolFeatures(VhostError)

Failed to set protocol features.

Tuple Fields of SetProtocolFeatures

0: VhostError

SetVringAddr(VhostError)

Failed to set vring address.

Tuple Fields of SetVringAddr

0: VhostError

SetVringBase(VhostError)

Failed to set vring base offset.

Tuple Fields of SetVringBase

0: VhostError

SetVringCall(VhostError)

Failed to set eventfd to signal used vring buffers.

Tuple Fields of SetVringCall

0: VhostError

SetVringEnable(VhostError)

Failed to enable or disable vring.

Tuple Fields of SetVringEnable

0: VhostError

SetVringKick(VhostError)

Failed to set eventfd for adding buffers to vring.

Tuple Fields of SetVringKick

0: VhostError

SetVringNum(VhostError)

Failed to set the size of the queue.

Tuple Fields of SetVringNum

0: VhostError

SocketConnect(Error)

Failed to connect socket.

Tuple Fields of SocketConnect

0: Error

SocketConnectOnMasterCreate(VhostError)

Failed to create Master from a UDS path.

Tuple Fields of SocketConnectOnMasterCreate

0: VhostError

TagTooLong

The tag for the Fs device was too long to fit in the config space.

Fields of TagTooLong

len: usizemax: usize