Enum devices::virtio::video::command::VideoCmd

pub enum VideoCmd {

    QueryCapability {
        queue_type: QueueType,
    },
    StreamCreate {
        stream_id: u32,
        coded_format: Format,
        input_resource_type: ResourceType,
        output_resource_type: ResourceType,
    },
    StreamDestroy {
        stream_id: u32,
    },
    StreamDrain {
        stream_id: u32,
    },
    ResourceCreate {
        stream_id: u32,
        queue_type: QueueType,
        resource_id: u32,
        plane_offsets: Vec<u32>,
        plane_entries: Vec<Vec<UnresolvedResourceEntry>>,
    },
    ResourceQueue {
        stream_id: u32,
        queue_type: QueueType,
        resource_id: u32,
        timestamp: u64,
        data_sizes: Vec<u32>,
    },
    ResourceDestroyAll {
        stream_id: u32,
        queue_type: QueueType,
    },
    QueueClear {
        stream_id: u32,
        queue_type: QueueType,
    },
    GetParams {
        stream_id: u32,
        queue_type: QueueType,
        is_ext: bool,
    },
    SetParams {
        stream_id: u32,
        queue_type: QueueType,
        params: Params,
        is_ext: bool,
    },
    QueryControl {
        query_ctrl_type: QueryCtrlType,
    },
    GetControl {
        stream_id: u32,
        ctrl_type: CtrlType,
    },
    SetControl {
        stream_id: u32,
        ctrl_val: CtrlVal,
    },

}

Variants

QueryCapability

Fields

queue_type: QueueType

StreamCreate

Fields

stream_id: u32

coded_format: Format

input_resource_type: ResourceType

output_resource_type: ResourceType

StreamDestroy

Fields

stream_id: u32

StreamDrain

Fields

stream_id: u32

ResourceCreate

Fields

stream_id: u32

queue_type: QueueType

resource_id: u32

plane_offsets: Vec<u32>

plane_entries: Vec<Vec<UnresolvedResourceEntry>>

The outer vector contains one entry per memory plane, whereas the inner vector contains all the memory entries that make a single plane (i.e. one for virtio objects, one or more for guest pages).

ResourceQueue

Fields

stream_id: u32

queue_type: QueueType

resource_id: u32

timestamp: u64

data_sizes: Vec<u32>

ResourceDestroyAll

Fields

stream_id: u32

queue_type: QueueType

QueueClear

Fields

stream_id: u32

queue_type: QueueType

GetParams

Fields

stream_id: u32

queue_type: QueueType

is_ext: bool

true if this command has been created from the GET_PARAMS_EXT guest command.

SetParams

Fields

stream_id: u32

queue_type: QueueType

params: Params

is_ext: bool

true if this command has been created from the SET_PARAMS_EXT guest command.

QueryControl

Fields

query_ctrl_type: QueryCtrlType

GetControl

Fields

stream_id: u32

ctrl_type: CtrlType

SetControl

Fields

stream_id: u32

ctrl_val: CtrlVal

Implementations

impl<'a> VideoCmd

pub fn from_reader(r: &'a mut Reader) -> Result<Self, ReadCmdError>

Reads a request on virtqueue and construct a VideoCmd value.