Module devices::virtio::wl

This module implements the virtio wayland used by the guest to access the host’s wayland server.

The virtio wayland protocol is done over two queues: in and out. The in queue is used for sending commands to the guest that are generated by the host, usually messages from the wayland server. The out queue is for commands from the guest, usually requests to allocate shared memory, open a wayland server connection, or send data over an existing connection.

Each WlVfd represents one virtual file descriptor created by either the guest or the host. Virtual file descriptors contain actual file descriptors, either a shared memory file descriptor or a unix domain socket to the wayland server. In the shared memory case, there is also an associated slot that indicates which hypervisor memory slot the memory is installed into, as well as a page frame number that the guest can access the memory from.

The types starting with Ctrl are structures representing the virtio wayland protocol “on the wire.” They are decoded and executed in the execute function and encoded as some variant of WlResp for responses.

There is one WlState instance that contains every known vfd and the current state of in queue. The in queue requires extra state to buffer messages to the guest in case the in queue is already full. The WlState also has a control socket necessary to fulfill certain requests, such as those registering guest memory.

The Worker is responsible for the poll loop over all possible events, encoding/decoding from the virtio queue, and routing messages in and out of WlState. Possible events include the kill event, available descriptors on the in or out queue, and incoming data on any vfd’s socket.

Structs

CtrlHeader

CtrlVfd

CtrlVfdDmabufSync

CtrlVfdNew

CtrlVfdNewCtxNamed

CtrlVfdNewDmabuf

CtrlVfdRecv

CtrlVfdSend

CtrlVfdSendVfd

CtrlVfdSendVfdV2

DescriptorsExhausted

GpuMemoryDesc

GpuMemoryPlaneDesc

VmRequester

VmRequesterState

Wl

WlState

WlVfd

Worker

dma_buf_sync

sync_file_info

Enums

WlError

WlRecv

WlResp

Constants

DMA_BUF_IOCTL_BASE

IN_BUFFER_LEN

NEXT_VFD_ID_BASE

QUEUE_SIZE

QUEUE_SIZES

VFD_ID_HOST_MASK

VIRTIO_WL_CMD_VFD_CLOSE

VIRTIO_WL_CMD_VFD_DMABUF_SYNC

VIRTIO_WL_CMD_VFD_HUP

VIRTIO_WL_CMD_VFD_NEW

VIRTIO_WL_CMD_VFD_NEW_CTX

VIRTIO_WL_CMD_VFD_NEW_CTX_NAMED

VIRTIO_WL_CMD_VFD_NEW_DMABUF

VIRTIO_WL_CMD_VFD_NEW_PIPE

VIRTIO_WL_CMD_VFD_RECV

VIRTIO_WL_CMD_VFD_SEND

VIRTIO_WL_CMD_VFD_SEND_FOREIGN_ID

VIRTIO_WL_CTRL_VFD_SEND_KIND_LOCAL

VIRTIO_WL_CTRL_VFD_SEND_KIND_VIRTGPU

VIRTIO_WL_CTRL_VFD_SEND_KIND_VIRTGPU_FENCE

VIRTIO_WL_CTRL_VFD_SEND_KIND_VIRTGPU_SIGNALED_FENCE

VIRTIO_WL_F_SEND_FENCES

VIRTIO_WL_F_TRANS_FLAGS

VIRTIO_WL_F_USE_SHMEM

VIRTIO_WL_PFN_SHIFT

VIRTIO_WL_RESP_ERR

VIRTIO_WL_RESP_INVALID_CMD

VIRTIO_WL_RESP_INVALID_FLAGS

VIRTIO_WL_RESP_INVALID_ID

VIRTIO_WL_RESP_INVALID_TYPE

VIRTIO_WL_RESP_OK

VIRTIO_WL_RESP_OUT_OF_MEMORY

VIRTIO_WL_RESP_VFD_NEW

VIRTIO_WL_RESP_VFD_NEW_DMABUF

VIRTIO_WL_VFD_CONTROL

VIRTIO_WL_VFD_DMABUF_SYNC_VALID_FLAG_MASK

VIRTIO_WL_VFD_FENCE

VIRTIO_WL_VFD_MAP

VIRTIO_WL_VFD_READ

VIRTIO_WL_VFD_WRITE

VIRTWL_SEND_MAX_ALLOCS

WL_SHMEM_ID

WL_SHMEM_SIZE

Functions

DMA_BUF_IOCTL_SYNC

SYNC_IOC_FILE_INFO

encode_resp

encode_vfd_hup

encode_vfd_new

encode_vfd_new_dmabuf

encode_vfd_recv

is_fence

process_in_queue

Handle incoming events and forward them to the VM over the input queue.

process_out_queue

Handle messages from the output queue and forward them to the display sever, if necessary.

Type Definitions

WlResult

Unions

CtrlVfdSendVfdV2Payload